% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/appendString.R
\name{appendString}
\alias{appendString}
\title{Append a string to another, using a separator, without appending a duplicate.}
\usage{
appendString(orig, app, sep = "\\n")
}
\arguments{
\item{orig}{- original string to append to}

\item{app}{- string to append}

\item{sep}{- string used to separate the original and appended strings. (default is newline)}
}
\description{
Function to append one string ('add') to another ('orig'), using a separator, without appending a duplicate.
}
\details{


The following examples apply (taking sep='/'):\cr
1: orig = '',     app = 'app',  result = 'app'\cr
2: orig = 'orig', app = 'app',  result = 'orig/app'\cr
3: orig = 'orig', app = 'orig', result = 'orig'
}

