% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcLogPrior.R
\name{calcLogPrior}
\alias{calcLogPrior}
\title{Calculate log-prior distribution based on input pdfType list}
\usage{
calcLogPrior(pdfType, v = NULL, x = NULL, lci = 0.05, uci = 0.95)
}
\arguments{
\item{pdfType}{- the pdfType list}

\item{v}{- values at which to calculate the prior values}

\item{x}{- vector of x coordinates used to plot the resulting distribution as a color image}

\item{lci}{- lower confidence interval for plotting the prior (if v is NULL)}

\item{uci}{- upper confidence interval for plotting the prior (if v is NULL)}
}
\value{
see details
}
\description{
Function to calculate log-prior distribution for plotting based on an input pdfType list.
}
\details{


If x is NULL, the function returns values of the log-prior distribution function
calculated using v (or coordinates based on the confidence interval). The names of the vector
provide the values of v at which the log prior was calculated. If x is given, the function returns
a list object of class 'pdfRects', with list elements lp, xl, xr, vl, and vr.
}

