% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compareModels.FisherySelFcns.R
\name{compareModels.FisherySelFcns}
\alias{compareModels.FisherySelFcns}
\title{Compare fishery selectivity functions from TCSAM2015 and rsimTCSAM model runs.}
\usage{
compareModels.FisherySelFcns(tcsam = NULL, rsim = NULL, years = list(TCF =
  2013), cast = "x", faceting = "y~x", showPlot = TRUE, pdf = NULL,
  width = 14, height = 8)
}
\arguments{
\item{tcsam}{- single TCSAM2015 model results object, or named list of such}

\item{rsim}{- single rsimTCSAM results object, or named list of such}

\item{years}{- list, by fishery name, of years to plot}

\item{cast}{- casting formula for excluding x,m,s factor levels from an average-at-size across unspecified factors}

\item{faceting}{- faceting formula involving y and factors in the cast}

\item{showPlot}{- flag to show/print plots immediately}

\item{pdf}{- name of pdf file to record plot output to}

\item{width}{- pdf page width (in inches)}

\item{height}{- pdf page width (in inches)}
}
\value{
list of ggplot2 objects
}
\description{
Function to compare fishery selectivity functions from TCSAM2015 and rsimTCSAM model runs.
}
\details{

}

