% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.Fits.AggCatchData.R
\name{compareModels.Fits.AggCatchData}
\alias{compareModels.Fits.AggCatchData}
\title{Compare fits to aggregated catch data (abundance or biomass) for TCSAM2015 model runs}
\usage{
compareModels.Fits.AggCatchData(mdfr, fleets = NULL, catch.types = NULL,
  logscale = FALSE, xlab = "Year", ylab = "", title = "", xlim = NULL,
  ylim = NULL, ggtheme = ggplot2::theme_grey(), showPlot = TRUE,
  pdf = NULL, width = 8, height = 6)
}
\arguments{
\item{mdfr}{- dataframe from call to \code{getMDFR.FitsForFleets(...)}}

\item{fleets}{- fleets to plot [NULL for all]}

\item{catch.types}{- catch types ('index.catch', retained.catch', etc) to plot [NULL for all]}

\item{logscale}{- flag (T/F) to plot on log scale}

\item{xlab}{- x axis label}

\item{ylab}{- y axis label}

\item{title}{- plot title}

\item{xlim}{- x axis limits}

\item{ylim}{- y axis limits}

\item{ggtheme}{- graphical theme for plot}

\item{showPlot}{- flag to show/print plots immediately}

\item{pdf}{- name of pdf file to record plot output to}

\item{width}{- pdf page width (in inches)}

\item{height}{- pdf page width (in inches)}
}
\value{
list of ggplot2 objects
}
\description{
Function to compare fits to aggregated catch data (abundance or biomass) for TCSAM2015 model runs.
}
\details{
Uses \code{reshape2::dcast(...)}
}

