% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.GrowthTransitionMatrices.R
\name{compareModels.GrowthTransitionMatrices}
\alias{compareModels.GrowthTransitionMatrices}
\title{Compare growth transition matrices for TCSAM2015 model runs}
\usage{
compareModels.GrowthTransitionMatrices(mdfr, nrow = 4, ncol = 3,
  xlab = "post-molt size (mm CW)", ylab = "proportion", xlim = NULL,
  ylim = NULL, ggtheme = ggplot2::theme_grey(), showPlot = TRUE,
  pdf = NULL, width = 8, height = 6, verbose = FALSE)
}
\arguments{
\item{mdfr}{- dataframe from call to \code{getMDFR.GrowthTransitionMatrices(...)} for n.at.z data}

\item{nrow}{- number of rows per page for output plots}

\item{ncol}{- number of columns per page for output plots}

\item{ggtheme}{- graphical theme for plot}

\item{showPlot}{- flag to show/print plots immediately}

\item{pdf}{- name of pdf file to record plot output to}

\item{width}{- pdf page width (in inches)}

\item{height}{- pdf page width (in inches)}

\item{verbose}{- flag (T/F) to print diagnostic output}
}
\value{
recursive list of lists of ggplot2 objects
}
\description{
Function to compare growth transition matrices for TCSAM2015 model runs.
}
\details{
Plots are produced by sex and faceted by parameter combination.
}

