% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.ObjFunValues.R
\name{compareModels.ObjFunValues}
\alias{compareModels.ObjFunValues}
\title{Plot all objective function values.}
\usage{
compareModels.ObjFunValues(repObj, variable = "objfun",
  ggtheme = ggplot2::theme_grey(), showPlot = TRUE, verbose = FALSE)
}
\arguments{
\item{repObj}{- model report object or list of report objects}

\item{variable}{- name of variable to plot}

\item{ggtheme}{- a ggplot2 theme}

\item{showPlot}{- flag to show plots immediately}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
list of ggplot2 objects corresponding to penalties, priors, and data plots.
}
\description{
Function to plot all objective function values.
}

