% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.ParamEsts.R
\name{compareModels.ParamEsts}
\alias{compareModels.ParamEsts}
\title{Function to compare parameter values from different models.}
\usage{
compareModels.ParamEsts(objs, dp = 0.01, fac = 2, nc = 3, nr = 4,
  showPlot = TRUE, pdf = "ModelComparisons.Params.pdf")
}
\arguments{
\item{objs}{- list of model results objects (each is a list with elements 'prsObj' and 'stdObj')}

\item{dp}{- percent difference between parameter value and upper/lower limits used to flag outliers}

\item{fac}{- number of std devs to extend uncertainty plots}

\item{nc}{- number of columns of plots per page}

\item{nr}{- number of rows of plots per page}

\item{showPlot}{- flag to show plots immediately}

\item{pdf}{- file name for printing plots to a pdf file (or NULL to print to screen)}
}
\value{
- list with dfr, vfr, and plots as elements
}
\description{
This function extracts and plots parameters values, together with their limits
(if any) and the posterior distributions implied by their estimated standard
errors from several models.
}

