% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.PopProcesses.R
\name{compareModels.PopProcesses}
\alias{compareModels.PopProcesses}
\title{Compare population process-related quantities from TCSAM2015 and rsimTCSAM model runs.}
\usage{
compareModels.PopProcesses(tcsams = NULL, rsims = NULL, showPlot = TRUE,
  pdf = NULL, width = 8, height = 6, verbose = FALSE)
}
\arguments{
\item{tcsams}{- single TCSAM2015 model report object, or named list of such}

\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{showPlot}{- flag to show/print plots immediately}

\item{pdf}{- name of pdf file to record plot output to}

\item{width}{- pdf page width (in inches)}

\item{height}{- pdf page width (in inches)}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
list of ggplot2 objects
}
\description{
Function to compare population process-related quantities from TCSAM2015 and rsimTCSAM model runs.
}
\details{
none.
}

