% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareModels.PopQuants.R
\name{compareModels.PopQuants}
\alias{compareModels.PopQuants}
\title{Compare population-related quantities from TCSAM2015 and rsimTCSAM model runs.}
\usage{
compareModels.PopQuants(tcsams = NULL, rsims = NULL, showPlot = TRUE,
  pdf = NULL, width = 8, height = 6)
}
\arguments{
\item{tcsams}{- single TCSAM2015 model results object, or named list of such}

\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{showPlot}{- flag to show/print plots immediately}

\item{pdf}{- name of pdf file to record plot output to}

\item{width}{- pdf page width (in inches)}

\item{height}{- pdf page width (in inches)}
}
\value{
list of ggplot2 objects
}
\description{
Function to compare population-related quantities from TCSAM2015 and rsimTCSAM model runs.
}
\details{

}

