% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.SelFcns.ByPC.R
\name{compareModels.SelFcns.ByPC}
\alias{compareModels.SelFcns.ByPC}
\title{Compare selectivity functions from TCSAM2015 and rsimTCSAM model runs.}
\usage{
compareModels.SelFcns.ByPC(tcsam = NULL, rsim = NULL, showPlot = TRUE,
  pdf = NULL, width = 8, height = 6)
}
\arguments{
\item{tcsam}{- single TCSAM2015 model results object, or named list of such}

\item{rsim}{- single rsimTCSAM results object, or named list of such}

\item{showPlot}{- flag to show/print plots immediately}

\item{pdf}{- name of pdf file to record plot output to}

\item{width}{- pdf page width (in inches)}

\item{height}{- pdf page width (in inches)}
}
\value{
list of ggplot2 objects
}
\description{
Function to compare selectivity functions from TCSAM2015 and rsimTCSAM model runs.
}

