% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.ZScores.AggCatchData.R
\name{compareModels.ZScores.AggCatchData}
\alias{compareModels.ZScores.AggCatchData}
\title{Compare zscores for fits to aggregated catch data (abundance or biomass) for TCSAM2015 model runs}
\usage{
compareModels.ZScores.AggCatchData(mdfr, fleets = NULL, catch.types = NULL,
  xlab = "Year", ylab = "", xlim = NULL, ylim = NULL,
  ggtheme = ggplot2::theme_grey(), showPlot = TRUE, pdf = NULL,
  width = 8, height = 6)
}
\arguments{
\item{mdfr}{- dataframe from call to \code{getMDFR.ZScoresForFleets(...)}}

\item{fleets}{- fleets to plot [NULL for all]}

\item{catch.types}{- catch types ('index.catch', retained.catch', etc) to plot [NULL for all]}

\item{xlab}{- x axis label}

\item{ylab}{- y axis label}

\item{xlim}{- x axis limits}

\item{ylim}{- y axis limits}

\item{ggtheme}{- graphical theme for plot}

\item{showPlot}{- flag to show/print plots immediately}

\item{pdf}{- name of pdf file to record plot output to}

\item{width}{- pdf page width (in inches)}

\item{height}{- pdf page width (in inches)}

\item{logscale}{- flag (T/F) to plot on log scale}
}
\value{
list of ggplot2 objects, named by fleet
}
\description{
Function to compare zscores for fits to aggregated catch data (abundance or biomass) for TCSAM2015 model runs.
}
\details{
Plots are produced by fleet and faceted by sex, maturity, shell condition, and catch type.
}

