% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.ZScores.NatZData.R
\name{compareModels.ZScores.NatZData}
\alias{compareModels.ZScores.NatZData}
\title{Compare z-scores from fits to size compositions from catch data for TCSAM2015 model runs}
\usage{
compareModels.ZScores.NatZData(mdfr, fleets = NULL, catch.types = NULL,
  variable = "pearsons", nrow = 4, ncol = 3, xlab = "Size (mm CW)",
  ylab = variable, xlim = NULL, ylim = NULL,
  ggtheme = ggplot2::theme_grey(), showPlot = TRUE, pdf = NULL,
  width = 8, height = 6, verbose = FALSE)
}
\arguments{
\item{mdfr}{- dataframe from call to \code{getMDFR.ZScoresForFleets(...)} for n.at.z data}

\item{fleets}{- fleets to plot [NULL for all]}

\item{catch.types}{- catch types ('index.catch', retained.catch', etc) to plot [NULL for all]}

\item{variable}{- variable to plot ("pearsons" or "nlls")}

\item{nrow}{- number of rows per page for output plots}

\item{ncol}{- number of columns per page for output plots}

\item{xlab}{- x axis label}

\item{ylab}{- y axis label}

\item{xlim}{- x axis limits}

\item{ylim}{- y axis limits}

\item{ggtheme}{- graphical theme for plot}

\item{showPlot}{- flag to show/print plots immediately}

\item{pdf}{- name of pdf file to record plot output to}

\item{width}{- pdf page width (in inches)}

\item{height}{- pdf page width (in inches)}

\item{verbose}{- flag (T/F) to print diagnostic output}
}
\value{
recursive list of lists  of ggplot2 objects, named by fleet, catch type, sex, maturity state, and shell condition
}
\description{
Function to compare z-scores from fits to size compositions from catch data for TCSAM2015 model runs.
}
\details{
Plots are produced by fleet, catch type, sex, maturity state, and shell condition and faceted by year.
}

