% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractModelResults.ParamsPlusStdDevs.R
\name{extractModelResults.ParamsPlusStdDevs}
\alias{extractModelResults.ParamsPlusStdDevs}
\title{Function to extract parameter estimates and standard deviations from different TCSAM2015 models}
\usage{
extractModelResults.ParamsPlusStdDevs(objs, dp = 0.01, verbose = FALSE)
}
\arguments{
\item{objs}{- list of TCSAM2015 model results objects (each is a list with elements 'prsObj' and 'stdObj')}

\item{dp}{- percent difference between parameter value and upper/lower limits used to flag outliers}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
- dataframe with the parameter estimates and uncertainty info.
}
\description{
This function extracts parameters values and standard deviations, together with their limits
(if any) from several TCSAM2015 models.
}
\details{
Returned parameter info dataframe has columns case, type, param, min, max, init, value, stdvv and scl.
}

