% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractModelResults.RepObjs.R
\name{extractModelResults.RepObjs}
\alias{extractModelResults.RepObjs}
\title{Extract (and optionally write to csv) model results from TCSAM2015 and rsimTCSAM report objects}
\usage{
extractModelResults.RepObjs(tcsams = NULL, rsims = NULL, path = NULL,
  cast.formula = NULL, label.value = "value", csv = NULL,
  verbose = FALSE)
}
\arguments{
\item{tcsams}{- single TCSAM2015 model report object, or named list of such}

\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{path}{- path to object to extract}

\item{cast.formula}{- casting formula for aggregation, or NULL for no aggregation}

\item{label.value}{- label for value column in output dataframe/csv}

\item{csv}{- csv file to write to (can be NULL)}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
A dataframe, written to csv if csv was not NULL.
}
\description{
Function to extract (and optionally write to csv) model results from TCSAM2015 and rsimTCSAM report objects.
}
\details{
If csv is NULL, then no file is written, but the created dataframe is still
returned. Uses \code{reshape2::dcast} to aggregate using the casting formula.
}

