% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareModels.ParamEsts.R
\name{extractModelResults.StdDevs}
\alias{extractModelResults.StdDevs}
\title{Function to extract parameter uncertainty (std) info from different models.}
\usage{
extractModelResults.StdDevs(objs, fac = 2)
}
\arguments{
\item{objs}{- model results objects (each is a list with elements 'stdObj')}

\item{fac}{- number of std devs to extend uncertainty plots}
}
\value{
- dataframe with uncertainty info
}
\description{
This function extracts the posterior distributions of parameters and
sdreport variables as implied by their estimated standard errors from several models.
}
\details{

}

