% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInputDataTimeframes.R
\name{getInputDataTimeframes}
\alias{getInputDataTimeframes}
\title{Get a dataframe describing the input data time frames}
\usage{
getInputDataTimeframes(tcsam, verbose = FALSE)
}
\arguments{
\item{tcsam}{- tcsam2015.rep list object}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
A dataframe.
}
\description{
A function to get a dataframe that describes the input data time frames.
}
\details{
returns a dataframe with columns
\itemize{
 \item year
 \item model
 \item fleet type
 \item fleet
 \item catch type
 \item data type
}
}

