% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMCMCFile}
\alias{getMCMCFile}
\title{Get a TCSAM2015 MCMC results file}
\usage{
getMCMCFile()
}
\value{
path to selected TCSAM2015 MCMC results file, or NULL (if user canceled selection).
}
\description{
Allows the user to select a TCSAM2015 MCMC results file using a file dialog.
}
\details{

}

