% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMCMCResults.R
\name{getMCMCResults}
\alias{getMCMCResults}
\title{Read a TCSAM 2015 MCMC (.R) file.}
\usage{
getMCMCResults(mcmcfile = NULL)
}
\arguments{
\item{mcmcfile}{- results file from a TCSAM2015 MCMC run to source. can be NULL.}
}
\value{
MCMC object as list. The returned list will be of class 'tcsam2015mcmc'.
}
\description{
Function to read a TCSAM 2015 MCMC (.R) file.
}
\details{
If resfile is NULL, the user will be prompted to identify a 
TCSAM 2015 MCMC file from which to source the results list object.
}

