% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.FitsForABData.R
\name{getMDFR.FitsForABData}
\alias{getMDFR.FitsForABData}
\title{Get fits to abundance or biomass data as a dataframe}
\usage{
getMDFR.FitsForABData(afits, ci = 0.95, verbose = FALSE)
}
\arguments{
\item{afits}{-}

\item{ci}{- confidence intervals}

\item{verbose}{- flag (T/F) to print dagnostic info}
}
\value{
dataframe
}
\description{
Function to get fits to abundance or biomass data as a dataframe.
}
\details{
none.
}

