% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.FitsForFleets.R
\name{getMDFR.FitsForFleets}
\alias{getMDFR.FitsForFleets}
\title{Get model fits to abundance, biomass or size frequencies for fleet data components}
\usage{
getMDFR.FitsForFleets(repObjs, fleet.type = "fishery",
  data.type = "abundance", verbose = FALSE)
}
\arguments{
\item{repObjs}{- single model report list object, or named list of them}

\item{fleet.type}{- fleet type ('fishery' or 'survey')}

\item{data.type}{- data type ('abundance', 'biomass', or 'n.at.z')}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
dataframe
}
\details{
Uses \code{getMDFR.FitsForABData()} and \code{getMDFR.FitsForSizeComps()}.
Returned dataframe has columns:
\itemize{
 \item{x - sex}
 \item{m - maturity}
 \item{s - shell condition}
 \item{y - year}
 \item{z - size bin [if data.type was 'n.at.z']}
 \item{val - value}
 \item{var - variable type}
 \item{data.type - data type ('abundance','biomass', or 'n.at.z')}
 \item{catch.type - catch type ('index.catch','retained.catch', etc.)}
 \item{fleet - fleet name}
 \item{fleet.type - fleet type ('survey' or 'fishery')}
 \item{model - model name}
}
Note: the 'z' column will be present only if the data.type requested was 'n.at.z'.
}

