% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.FitsForSizeComps.R
\name{getMDFR.FitsForSizeComps}
\alias{getMDFR.FitsForSizeComps}
\title{Get fits for size comps as a dataframe}
\usage{
getMDFR.FitsForSizeComps(fits, mc, verbose = FALSE)
}
\arguments{
\item{fits}{- list of fits for size compositions}

\item{mc}{- model configuration list}

\item{verbose}{- flag (T/F) to print dagnostic info}
}
\value{
dataframe
}
\description{
Get fits for size comps as a dataframe.
}
\details{
Uses \code{reshape2::melt(...)}. 
Returned dataframe has columns:
\itemize{
 \item{x - sex}
 \item{m - maturity}
 \item{s - shell condition}
 \item{y - year}
 \item{z - size bin}
 \item{val - value}
 \item{var - variable type}
}
}

