% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Special.R
\name{getMDFR.MeanGrowthIncrements}
\alias{getMDFR.MeanGrowthIncrements}
\title{Get mean growth increments from model results from TCSAM2015 and rsimTCSAM model runs as a dataframe}
\usage{
getMDFR.MeanGrowthIncrements(tcsams, rsims, verbose = FALSE)
}
\arguments{
\item{tcsams}{- single TCSAM2015 model report object, or named list of such}

\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
nothing
}
\description{
Function to get mean growth increments from model results from TCSAM2015 and rsimTCSAM model runs as a dataframe.
}
\details{
Extracts mean growth increments.
}

