% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Special.R
\name{getMDFR.NatMort}
\alias{getMDFR.NatMort}
\title{Get natural mortality rates from model results from TCSAM2015 and rsimTCSAM model runs as a dataframe}
\usage{
getMDFR.NatMort(tcsams = NULL, rsims = NULL, type = c("M_cxm", "M_yxm",
  "M_yxmsz"), verbose = FALSE)
}
\arguments{
\item{tcsams}{- single TCSAM2015 model report object, or named list of such}

\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{type}{- flag indicating which M's to extract ('M_cxm', 'M_yxm' or 'M_yxmsz')}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
dataframe with columns
\itemize{ 
 \item{'M_cxm:',  'modeltype', 'model', 'pc', 'x', 'm', 'val'}
 \item{'M_yxm:',  'modeltype', 'model',  'y', 'x', 'm', 'val'}
 \item{'M_yxmsz:','modeltype', 'model', 'y', 'x', 'm', 's', 'z', 'val'}
}
}
\description{
Function to get natural mortality rates from model results from TCSAM2015 and rsimTCSAM model runs as a dataframe.
}
\details{
Extracts natural mortality rates
}

