% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.ZScoresForABData.R
\name{getMDFR.ZScoresForABData}
\alias{getMDFR.ZScoresForABData}
\title{Get z-scores for abundance, biomass data from fits to data as dataframe}
\usage{
getMDFR.ZScoresForABData(afits, verbose = FALSE)
}
\arguments{
\item{afits}{-}

\item{verbose}{- flag (T/F) to print dagnostic info}
}
\value{
dataframe
}
\description{
Function to get z-scores for abundance, biomass data from fits to data as dataframe.
}

