% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.ZScoresForCatchData.R
\name{getMDFR.ZScoresForCatchData}
\alias{getMDFR.ZScoresForCatchData}
\title{Get fits to catch data (abundance, biomass) as z-scores in a dataframe}
\usage{
getMDFR.ZScoresForCatchData(repObjs, fleet.type = "fishery",
  verbose = FALSE)
}
\arguments{
\item{verbose}{- flag (T/F) to print diagnostic info}

\item{fit}{- fits list for given catch type}
}
\value{
dataframe
}
\description{
Function to get fits to catch data (abundance, biomass) as z-scores in a dataframe.
}
\details{
None.
}

