% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.ZScoresForSizeComps.R
\name{getMDFR.ZScoresForSizeComps}
\alias{getMDFR.ZScoresForSizeComps}
\title{Get residuals from size frequency fits as dataframe}
\usage{
getMDFR.ZScoresForSizeComps(fits, mc, verbose = FALSE)
}
\arguments{
\item{fits}{- list of fits}

\item{mc}{- model configuration list}

\item{verbose}{- flag (T/F) to print dagnostic info}
}
\value{
dataframe
}
\description{
Get Pearson's residuals and negative log-likelihood components
from size frequency fits as dataframe.
}
\details{
Uses \code{reshape2::melt(...)}. 
Returned dataframe has columns:
\itemize{
 \item{x - sex}
 \item{m - maturity}
 \item{s - shell condition}
 \item{y - year}
 \item{z - size bin}
 \item{val - value}
 \item{var - variable type}
}
}

