% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getModelResults.R
\name{getModelResults}
\alias{getModelResults}
\title{Read a TCSAM 2015 model report (.rep) file.}
\usage{
getModelResults(resfile = NULL)
}
\arguments{
\item{resfile}{- results file from a TCSAM2014 model run to source. can be NULL.}
}
\value{
results model object as list. The returned list will be of class 'tcsam2015'.
}
\description{
Function to read a TCSAM 2015 report file (.rep).
}
\details{


If resfile is NULL, the user will be prompted to identify a
TCSAM 2015 model report file from which to source the results list object.
}

