% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.R
\name{getObj}
\alias{getObj}
\title{Get an object from a list object using the specified path into the list.}
\usage{
getObj(path, lst)
}
\arguments{
\item{path}{- path in the list to the desired object (using '/' as separator for list levels)}

\item{lst}{- the list to extract the object from}
}
\value{
the specified object
}
\description{
Function to get an object from a list object using the specified path into the list.
}

