% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObjFunValues.AggregateCatch.R
\name{getObjFunValues.AggregateCatch}
\alias{getObjFunValues.AggregateCatch}
\title{Get aggregate catch components in the objective function as a melted dataframe.}
\usage{
getObjFunValues.AggregateCatch(aggC, data.type = "abundance",
  verbose = FALSE)
}
\arguments{
\item{aggC}{- a list object reflecting an aggregate catch data source's contributions to the objective function}

\item{data.type}{- data type (abundance or biomass) of aggregated catch}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
a melted dataframe
}
\description{
Function to get aggregate catch components in the objective function as a melted dataframe.
}
\details{
The returned dataframe has columns named 
"data.type", "fit.type", "nll.type", "year" (a dummy value = -1), 
"sex", "maturity", "shell_condition", "variable", and "value".

The "variable" column indicates whether the "value" is a weight ('wgt'),
negative log-likelihood ('nll'), or objective function value ('objfun').
}

