% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObjFunValues.CatchData.R
\name{getObjFunValues.CatchData}
\alias{getObjFunValues.CatchData}
\title{Get catch data components in the objective function as a melted dataframe.}
\usage{
getObjFunValues.CatchData(catchData, catch.type = "total.catch",
  verbose = FALSE)
}
\arguments{
\item{catchData}{- a list object reflecting a catch data source's contributions to the objective function}

\item{catch.type}{- type of catch data ('total.catch','retained.catch','discard.catch')}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
a melted dataframe
}
\description{
Function to get catch data components in the objective function as a melted dataframe.
}
\details{
The returned dataframe has columns named 
"catch.type", "data.type",      
"fit.type", "nll.type", "year", "sex",
"maturity", "shell_condition", "variable", "value"

The "variable" column indicates whether the "value" is a weight ('wgt'),
negative log-likelihood ('nll'), or objective function value ('objfun').
}

