% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObjFunValues.Data.R
\name{getObjFunValues.Data}
\alias{getObjFunValues.Data}
\title{Get data components in the objective function as a melted dataframe.}
\usage{
getObjFunValues.Data(repObjs, mdl = NULL, verbose = FALSE)
}
\arguments{
\item{repObjs}{- tcsam2015 model results object or list of such}

\item{mdl}{- name to associate with model results object}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
a melted dataframe
}
\description{
Function to get data (fisheries, surveys) components in the objective function as a melted dataframe.
}
\details{
If repObjs is a list of tcsam2015 model results objects, then the function
is called recursively for each object, with the associated list component name used as 
mdl. If repObjs is a tcsam2015 model results object and mdl is NULL (the default), then 
repObjs$mc$configName is used as the model name.

The returned dataframe has columns named 
"model", "source.type", "source.name", "catch.type",  "data.type",      
"fit.type", "nll.type", "year", "sex",
"maturity", "shell_condition", "variable", "value"

The "variable" column indicates whether the "value" is a weight ('wgt'),
negative log-likelihood ('nll'), or objective function value ('objfun').
}

