% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getObjFunValues.NatZ.R
\name{getObjFunValues.NatZ}
\alias{getObjFunValues.NatZ}
\title{Get size comps components in the objective function as a melted dataframe.}
\usage{
getObjFunValues.NatZ(nAtZ)
}
\arguments{
\item{nAtZ}{- a list object reflecting a size comps' contributions to the objective function}
}
\value{
a melted dataframe
}
\description{
Function to get size comps components in the objective function as a melted dataframe.
}
\details{


The returned dataframe has columns named
"data.type", "fit.type", "nll.type", "year",
"sex", "maturity", "shell_condition", "variable", and "value".

The "variable" column indicates whether the "value" is a weight ('wgt'),
negative log-likelihood ('nll'), objective function value ('objfun'), input
sample size ('ss') or effective sample size ('effN').
}

