% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getObjFunValues.Penalties.R
\name{getObjFunValues.Penalties}
\alias{getObjFunValues.Penalties}
\title{Get model penalties in the objective function as a melted dataframe.}
\usage{
getObjFunValues.Penalties(repObj, mdl = NULL)
}
\arguments{
\item{repObj}{- tcsam2015 model results object or list of such}

\item{mdl}{- name to associate with model results object}
}
\value{
a melted dataframe
}
\description{
Function to get model penalties as a melted dataframe.
}
\details{


If repObj is a list of tcsam2015 model results objects, then the function
is called recursively for each object, with the associated list component name used as
mdl. If repObj is a tcsam2015 model results object and mdl is NULL (the default), then
repObj$mc$configName is used as the model name.

The returned dataframe has columns named
"model", "type", "category", "name", "level", "variable", and "value".

The "variable" column indicates whether the "value" is a weight ('wgt'),
negative log-likelihood ('nll'), or objective function value ('objfun').
}

