% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObjFunValues.Priors.R
\name{getObjFunValues.Priors}
\alias{getObjFunValues.Priors}
\title{Get model priors in the objective function as a melted dataframe.}
\usage{
getObjFunValues.Priors(res, mdl = NULL, verbose = FALSE)
}
\arguments{
\item{res}{- tcsam2015 model results object or list of such}

\item{mdl}{- name to associate with model results object}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
a melted dataframe
}
\description{
Function to get model priors as a melted dataframe.
}
\details{
If res is a list of tcsam2015 model results objects, then the function
is called recursively for each object, with the associated list component name used as 
mdl. If res is a tcsam2015 model results object and mdl is NULL (the default), then 
res$mc$configName is used as the model name.

The returned dataframe has columns named 
"model", "type", "category", "name", "level", "variable", and "value".

The "variable" column indicates whether the "value" is a weight ('wgt'),
negative log-likelihood ('nll'), or objective function value ('objfun').
}

