% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObjFunValues.Surveys.R
\name{getObjFunValues.Surveys}
\alias{getObjFunValues.Surveys}
\title{Get survey components in the objective function as a melted dataframe.}
\usage{
getObjFunValues.Surveys(repObj, model = NULL)
}
\arguments{
\item{repObj}{- tcsam2015 model report object or list of such}

\item{model}{- name to associate with model report object}
}
\value{
a melted dataframe
}
\description{
Function to get survey components in the objective function as a melted dataframe.
}
\details{
If repObj is a list of tcsam2015 model report objects, then the function
is called recursively for each object, with the associated list component name used as the
model. If repObj is a tcsam2015 model report object and model is NULL (the default), then 
repObj$mc$configName is used as the model name.

The returned dataframe has columns named 
"model", "source.type", "source.name", "catch.type", "data.type",       
"fit.type", "nll.type", "year", "sex",
"maturity", "shell_condition", "variable", "value"

The "variable" column indicates whether the "value" is a weight ('wgt'),
negative log-likelihood ('nll'), or objective function value ('objfun').
}

