% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPlotSymbology}
\alias{getPlotSymbology}
\title{Get symbology info for plotting purposes}
\usage{
getPlotSymbology(res, col.SX = c("green", "blue"), pch.MS = c(21, 22),
  lty.SC = c(2, 1))
}
\arguments{
\item{res}{- results list from TCSAM2014 model run}
}
\value{
list with elements 'col','pch','lty', and 'leg.info'. \cr
leg.info is in turn a list, with elements \cr
'col','txt','txt.col','pch', and 'lty'.
}
\description{
Get a list with the TCSAM2014 standard plot symbology
}
\details{

}

