% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrs.R
\name{getPrs}
\alias{getPrs}
\title{Get a TCSAM2015 prs object by reading a parameters csv file}
\usage{
getPrs(csvFile)
}
\arguments{
\item{csvFile}{- parameters csv file from a TCSAM2015 model run. can be NULL.}
}
\value{
prs model object (a list). The returned object will be a list of class 'tcsam2015.prs'.
}
\description{
Function to get a TCSAM2015 prs object by reading a parameters csv file.
}
\details{
If \code{csvFile} is NULL, the user will be prompted to identify a 
TCSAM2015 model report file from which to source the results object.
The returned object will be a list of class 'tcsam2015.prs'.
}

