% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRep.R
\name{getRep}
\alias{getRep}
\title{Get a TCSAM2015 report object (by reading a TCSAM2015 model report [.rep] file)}
\usage{
getRep(repFile)
}
\arguments{
\item{repFile}{- report file from a TCSAM2015 model run to source. can be NULL.}
}
\value{
TCSAM2015 report object (a list). The returned object will be a list of class 'tcsam2015.rep'.
}
\description{
Function to get a TCSAM2015 report object
}
\details{
If \code{repFile} is NULL, the user will be prompted to identify a 
TCSAM2015 model report file from which to source the results object.
The returned object will be a list of class 'tcsam2015.rep'.
}

