% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getResultsFile.R
\name{getResultsFile}
\alias{getResultsFile}
\title{Get a TCSAM2015 model results file}
\usage{
getResultsFile()
}
\value{
path to selected TCSAM2015 model results file, or NULL (if user canceled selection).
}
\description{
Allows the user to select a TCSAM2015 model results file using a file dialog.
}
\details{

}

