% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRunCommands.R
\name{getRunCommands}
\alias{getRunCommands}
\title{Generate run commands for a tcsam2015 model run}
\usage{
getRunCommands(os = "osx", model = "tcsam2015", path2model = "",
  configFile = "", pin = FALSE, hess = FALSE, mcmc = FALSE,
  mc.N = 1e+06, mc.save = 1000, mc.scale = 1000, jitter = FALSE,
  jit.seed = NULL, cleanup = TRUE)
}
\arguments{
\item{os}{- 'win', 'mac' or 'osx'}

\item{model}{- admb model name}

\item{path2model}{- path to model}

\item{configFile}{- filename (including path) to model configuration file}

\item{pin}{- flag (T/F) to use a pin file}

\item{hess}{- flag (T/F) to calculate the hessian}

\item{mcmc}{- flag (T/F) to do mcmc calculations}

\item{mc.N}{- number of mcmc iterations to do}

\item{mc.save}{- number of iterations to skip when saving mcmc calculations}

\item{mc.scale}{- number of iterations to adjust scale for mcmc calculations}

\item{jitter}{- flag (T/F) to use jitter initial values}

\item{jit.seed}{- value for random number seed to generate jitter}

\item{cleanup}{- flag (T/F) to clean up unnecessary files}
}
\description{
Function to generate a script to make a tcsam2015 model run
}
\details{
. If \code{cleanup} is TRUE, then the executable, .bar, .b0*, .p0*, .r0*, variance, 
EchoOut.dat, CheckFile.dat, and fimn.log files are deleted. If not, then only the executable
is deleted.\cr
If the path associated with \code{configFile} is a relative one, it should
be relative to the \code{path} variable.
}

