% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStd.R
\name{getStd}
\alias{getStd}
\title{Get a TCSAM2015 std object by reading a .std file}
\usage{
getStd(stdFile)
}
\arguments{
\item{csvFile}{- parameters csv file from a TCSAM2015 model run. can be NULL.}
}
\value{
std model object (a list). The returned object will be a list of class 'tcsam2015.std'.
}
\description{
Function to get a TCSAM2015 std object by reading a .std file
}
\details{
If \code{csvFile} is NULL, the user will be prompted to identify a 
TCSAM2015 model report file from which to source the results object.
The returned object will be a list of class 'tcsam2015.std'.
}

