% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotAggregateCatchData}
\alias{plotAggregateCatchData}
\title{Plot aggregate catch data}
\usage{
plotAggregateCatchData(name = NULL, obs = NULL, mod = NULL,
  logscale = FALSE, pch = c(21, 22, 23), lty = c(2, 1, 1), lwd = c(2, 2,
  2), col = c("green", "blue", "black"), xlab = "year", ylab = "",
  xlim = NULL, ylim = NULL, plot.legend = TRUE)
}
\arguments{
\item{name}{- fishery or survey name}

\item{obs}{- list of observed data}

\item{mod}{- list of model-predicted data}

\item{logscale}{- flag (T/F) to plot on ln-scale}

\item{pch}{- vector of point types to use (for observed data)}

\item{lty}{- vector of line types to use (for model-predicted data)}

\item{lwd}{- vector of line widths}

\item{col}{- vector of colors to use for points and lines}

\item{xlab}{- x axis title}

\item{ylab}{- y axis title}

\item{xlim}{- x axis limits}

\item{ylim}{- y axis limits}

\item{plot.legend}{- flag (T/F) to include a legend}
}
\description{
Plot aggregate (numbers or biomass) catch data from fisheries or surveys
}
\details{

}

