% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotAggregateCatchDataGG.R
\name{plotAggregateCatchDataGG}
\alias{plotAggregateCatchDataGG}
\title{Plot aggregate catch data}
\usage{
plotAggregateCatchDataGG(label = NULL, obs = NULL, mod = NULL,
  ci = 0.95, pdfType = "normal", logscale = FALSE, xlab = "year",
  ylab = "", xlim = NULL, ylim = NULL, ggtheme = theme_grey(),
  showPlot = FALSE)
}
\arguments{
\item{label}{- fishery or survey name}

\item{obs}{- list of observed data}

\item{mod}{- list of model-predicted data}

\item{ci}{- confidence interval}

\item{pdfType}{- pdf type for ci calculations}

\item{logscale}{- flag to plot on log scale}

\item{xlab}{- x axis title}

\item{ylab}{- y axis title}

\item{xlim}{- x axis limits}

\item{ylim}{- y axis limits}

\item{ggtheme}{- ggplot2 theme}

\item{showPlot}{- flag to print plot immediately}
}
\value{
ggplot2 object
}
\description{
Plot aggregate (numbers or biomass) catch data from fisheries or surveys
}
\details{

}

