% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotBars.R
\name{plotBars}
\alias{plotBars}
\title{Create or add a bar plot.}
\usage{
plotBars(y = NULL, x = NULL, w = NULL, df = NULL, xCol = NULL,
  yCol = NULL, xloc = "L", xlabs = NA, overplot = FALSE,
  stack = FALSE, barwidth = 100, groupwidth = 100, barclr = NULL,
  ymin = 0, plotLegend = TRUE, leg.loc = "topright", leg.text = "", ...)
}
\arguments{
\item{y}{- matrix of values for bars (group elements as rows, series as columns)}

\item{x}{- x coordinate for bars (groups) (optional)}

\item{w}{- bar (group) width (or widths, if a vector) (optional)}

\item{df}{- dataframe (optional way to specify x, y)}

\item{xCol}{- column name in df to use as x}

\item{yCol}{- vector of column names to use as y}

\item{xloc}{- type of x location specified by x."L" - left cutpoint, "C" - center point, "R" - right cutpoint}

\item{overplot}{- TRUE to plot bars on last graph}

\item{stack}{- TRUE to stack bars}

\item{barclr}{- vector of bar colors by group}

\item{ymin}{- min y to plot}

\item{plotLegend}{- TRUE/FALSE flag to plot legend}

\item{leg.loc}{- location for legend (e.g., "topright")}

\item{leg.text}{- text labels for legend}

\item{bar}{width   - scale (in percent) to use for bar widths}

\item{groupwidths}{- scale (in percent) to use for group widths}
}
\description{
Function to create or add to a bar plot.
}
\details{

}

