% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCompsAsCirclesGG.R
\name{plotCompsAsCirclesGG}
\alias{plotCompsAsCirclesGG}
\title{Create or add to a circle plot of age or size comp.s.}
\usage{
plotCompsAsCirclesGG(dfr, x = NULL, y = NULL, z = NULL, category = NULL,
  faceting = NULL, title = NA, xlim = NULL, ylim = NULL, xlab = NA,
  ylab = NA, alpha = 0.6, ggtheme = theme_grey(), showPlot = TRUE,
  verbose = FALSE)
}
\arguments{
\item{dfr}{- dataframe from which to extract columns}

\item{x}{- column name with x coordinates of circle centers}

\item{y}{- column name with y coordinates of circle centers}

\item{z}{- column name with values to be plotted as circles}

\item{category}{- column name with values used as categories for plot}

\item{faceting}{- formula (as string) for faceting (NULL=no faceting)}

\item{title}{- title for plot}

\item{xlim}{- x axis limits}

\item{ylim}{- y axis limits}

\item{xlab}{- x axis label}

\item{ylab}{- y axis label}

\item{alpha}{- alpha level for transparency (0-1)}

\item{ggtheme}{- ggplot2 theme}

\item{showPlot}{- flag to show (print) plot immediately on current graphics device}

\item{verbose}{- flag (T/F) to print dagnostic info}
}
\value{
ggplot2 object
}
\description{
Creates or adds to a circle plot of age or size comp.s.
}

