% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotDataModelComparisons}
\alias{plotDataModelComparisons}
\title{Plot model-simulated and observed data for comparison.}
\usage{
plotDataModelComparisons(res)
}
\arguments{
\item{res}{- results list object from report file}
}
\description{

}

