% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotEffNsForSizeFreqs}
\alias{plotEffNsForSizeFreqs}
\title{Plot ISSs and ESSs from size frequency fits as time series}
\usage{
plotEffNsForSizeFreqs(fits, mc, sxs = c(mc$SXs, "ALL_SEX"), mss = c(mc$MSs,
  "ALL_MATURITY"), scs = c(mc$SCs, "ALL_SHELL_CONDITION"), label = "")
}
\arguments{
\item{fits}{- list of fits}

\item{mc}{- model configuration list}

\item{label}{- plot label}
}
\description{
Function to plot input and effective sample sizes (ISSs and ESSs) from size frequencies as time series.
}
\details{

}

