% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEffNsGG.R
\name{plotEffNsGG}
\alias{plotEffNsGG}
\title{Plot input and estimated sample sizes from size frequency fits}
\usage{
plotEffNsGG(fits, mc, sxs = c(mc$dims$x$nms, "ALL SEX"),
  mss = c(mc$dims$m$nms, "ALL MATURITY"), scs = c(mc$dims$s$nms,
  "ALL SHELL"), label = "", ggtheme = theme_grey(), showPlot = TRUE,
  verbose = FALSE)
}
\arguments{
\item{fits}{- list of fits}

\item{mc}{- model configuration list}

\item{sxs}{- character vector of sexes to plot}

\item{mss}{- character vector of maturity states to plot}

\item{scs}{- character vector of shell conditions to plot}

\item{label}{- plot label}

\item{ggtheme}{- ggplot2 theme}

\item{showPlot}{- flag to show (print) plot immediately on current graphics device}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
list of ggplot2 plot objects
}
\description{
Plot input and estimated sample sizes
from size frequency fits using ggplot2.
}

