% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotEffortRegressions.R
\name{plotEffortRegressions}
\alias{plotEffortRegressions}
\title{Plot regressions of capture rate on effort to assess effort extrapolations.}
\usage{
plotEffortRegressions(repObj, showPlot = TRUE)
}
\arguments{
\item{repObj}{- tcsam2015 report object}

\item{showPlot}{- flag to show plots immediately}
}
\value{
list of ggplot objects (by fishery name) showing regressions
}
\description{
Function to plot  regressions of capture rate on effort to assess effort extrapolations.
}
\details{

}

