% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotExploitationRatesGG.R
\name{plotExploitationRatesGG}
\alias{plotExploitationRatesGG}
\title{Plot time series of exploitation rates based on abundance (natural, discards, retained mortality) by fishery and sex.}
\usage{
plotExploitationRatesGG(res, showPlot = TRUE)
}
\arguments{
\item{res}{- results list from a TCSAM2015 model run}

\item{showPlots}{- flag (T/F) to show plots}
}
\value{
list of ggplot objects
}
\description{
Plot time series of exploitation rates based on abundance (natural, discards, retained mortality) by fishery and sex.
}
\details{

}

