% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotFisheriesResults}
\alias{plotFisheriesResults}
\title{Plot model results for fisheries.}
\usage{
plotFisheriesResults(res, showPlots = TRUE)
}
\arguments{
\item{res}{- results list from a TCSAM2015 model run}

\item{showPlots}{- flag (T/F) to show plots}
}
\value{
nested list of ggplot objects
}
\description{
Function to plot model results for fisheries.
}
\details{

}

