% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotFisheryCatchesGG}
\alias{plotFisheryCatchesGG}
\title{Plot time series of the fishery catch (numbers: captured, discards, retained) by fishery and sex.}
\usage{
plotFisheryCatchesGG(res, showPlots = TRUE)
}
\arguments{
\item{res}{- results list from a TCSAM2015 model run}

\item{showPlots}{- flag (T/F) to show plots}
}
\value{
list of ggplot objects
}
\description{
Plot time series of the fishery catch (numbers: captured, discards, retained) by fishery and sex.
}
\details{

}

