% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotFisheryYieldsGG.R
\name{plotFisheryYieldsGG}
\alias{plotFisheryYieldsGG}
\title{Plot time series of the fishery yield (biomass: captured, discards, retained) by fishery and sex.}
\usage{
plotFisheryYieldsGG(res, showPlots = TRUE)
}
\arguments{
\item{res}{- results list from a TCSAM2015 model run}

\item{showPlots}{- flag (T/F) to show plots}
}
\value{
list of ggplot objects
}
\description{
Plot time series of the fishery yield (biomass: captured, discards, retained) by fishery and sex.
}
\details{

}

