% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotFishingRates}
\alias{plotFishingRates}
\title{Plot time series of the fully-selected fishery capture rates.}
\usage{
plotFishingRates(res)
}
\arguments{
\item{res}{- results list from a TCSAM2014 model run}
}
\description{
Plot time series of the fully-selected fishery capture rates.
}
\details{

}

