% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotFitsGG.MeanSizeComps.R
\name{plotFitsGG.MeanSizeComps}
\alias{plotFitsGG.MeanSizeComps}
\title{Plot fits to mean size comps}
\usage{
plotFitsGG.MeanSizeComps(fits, mc, sxs = c(mc$dims$x$nms, "ALL_SEX"),
  mss = c(mc$dims$m$nms, "ALL_MATURITY"), scs = c(mc$dims$s$nms,
  "ALL_SHELL"), label = "", ggtheme = theme_grey(), showPlot = TRUE)
}
\arguments{
\item{fits}{- list of fits}

\item{mc}{- model configuration list}

\item{sxs}{- character vector of sexes to plot}

\item{mss}{- character vector of maturity states to plot}

\item{scs}{- character vector of shell conditions to plot}

\item{label}{- plot label}

\item{ggtheme}{- ggplot2 theme}

\item{showPlot}{- flag to show (print) plot immediately on current graphics device}
}
\value{
list of list of ggplot2 plot objects
}
\description{
Plot mean size comps fits using ggplot2.
}
\details{

}

