% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMatrix.R
\name{plotMatrix}
\alias{plotMatrix}
\title{Plot a matrix as a color map.}
\usage{
plotMatrix(mat, zlab = "", xlab = "", ylab = "", zmax = NULL,
  zmin = 0, title = NULL, cex = 1, colorscale = c("hot", "cold",
  "coldhot"), plotsize = c(970, 780), showDiagonal = TRUE)
}
\arguments{
\item{mat}{- matrix to plot}

\item{zlab}{- z label}

\item{xlab}{- label for x axis}

\item{ylab}{- label for y axis}

\item{zmax}{- max value for color scale}

\item{zmin}{- min value for color scale}

\item{title}{- title for plot}

\item{cex}{- character expansion factor for plot}

\item{colorscale}{- color scale to use for plot ('hot', 'cold' or 'coldhot')}

\item{plotsize}{- 2 element vector giving plot size in pixels (c(W,H))}

\item{showDiagonal}{- TRUE/FALSE to plot diagonal line}
}
\description{
Function to plot a matrix as a color map.
}

