% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotMCMCDensity}
\alias{plotMCMCDensity}
\title{Calculate and plot a posterior density based on an MCMC data vector.}
\usage{
plotMCMCDensity(data, scaleBy = 1, plotEst = TRUE, add = TRUE,
  clr = "cyan", alpha = 0.25, xlim = NULL, ylim = NULL, xlabel = "",
  label = "")
}
\arguments{
\item{data}{- the MCMC data from which to estimate the posterior density.}

\item{scaleBy}{- factor to scale data by}

\item{plotEst}{- flag (T/F) to plot the MLE estimate (assumed to be 1st value)}

\item{add}{- flag (T/F) to add to existing plot (creates new plot if FALSE)}

\item{clr}{- color for density plot}

\item{alpha}{- transparency value to apply to clr}

\item{xlim}{- x axis limits (if add=FALSE)}

\item{ylim}{- y axis limits (if add=FALSE)}

\item{xlabel}{- label for x axis (if add=FALSE)}

\item{label}{- label for plot (if add=FALSE)}
}
\description{
Function to calculate and plot a posterior density based on an MCMC data vector.
}
\details{

}

