% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMCMCResults.forArray3D.R
\name{plotMCMCResults.forArray3D}
\alias{plotMCMCResults.forArray3D}
\title{Calculate and plot a posterior density based on a 3d MCMC data aray.}
\usage{
plotMCMCResults.forArray3D(data, scaleBy = 1, doPlot = TRUE,
  plotEst = FALSE, add = TRUE, colorscale = c("coldhot", "hot", "cold",
  "jet", NULL), alpha = 0.25, xlim = NULL, ylim = NULL, xlabel = "",
  ylabel = "", label = "")
}
\arguments{
\item{data}{- the MCMC data array from which to estimate the posterior densities.}

\item{scaleBy}{- factor to scale data by}

\item{plotEst}{- flag (T/F) to plot the MLE estimate (assumed to be 1st value)}

\item{add}{- flag (T/F) to add to existing plot (creates new plot if FALSE)}

\item{colorscale}{- color scale to use for the density plot}

\item{alpha}{- transparency value to apply to the colorscale}

\item{xlim}{- x axis limits (if add=FALSE)}

\item{ylim}{- y axis limits (if add=FALSE)}

\item{xlabel}{- label for x axis (if add=FALSE)}

\item{label}{- label for plot (if add=FALSE)}
}
\description{
Function to calculate and plot posterior densities based on a 3d MCMC data array.
}
\details{
Uses functions
\itemize{
 \item wtsUtilities::createColorScale(...)
}
}

