% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotMCMCResults}
\alias{plotMCMCResults}
\title{Plot compiled MCMC results.}
\usage{
plotMCMCResults(lst)
}
\arguments{
\item{lst}{- list of the compiled MCMC results to plot}
}
\description{
Function to plot compiled MCMC results.
}
\details{

}

