% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotMDFR.Bars.R
\name{plotMDFR.Bars}
\alias{plotMDFR.Bars}
\title{Plot melted dataframe using a barchart format.}
\usage{
plotMDFR.Bars(mdfr, x = NULL, value.var = "val", agg.formula = NULL,
  agg.function = sum, ..., colour = NULL, fill = NULL, faceting = NULL,
  position = "dodge", xlab = "", ylab = "", units = "",
  lnscale = FALSE, title = "", guideTitleColour = NULL,
  guideTitleFill = NULL, showPlot = FALSE)
}
\arguments{
\item{mdfr}{- melted dataframe}

\item{x}{- column name for x-axis values}

\item{value.var}{- column name for values to aggregate (value.var in cast)/plot on y-axis}

\item{agg.formula}{- aggregation formula (left-hand side of cast formula)}

\item{agg.function}{- aggregation function (fun.aggregate in cast)}

\item{...}{- further arguments passed to aggregating function}

\item{colour}{- column name to which colour aesthetic is mapped}

\item{fill}{- column name to which fill aesthetic is mapped}

\item{faceting}{- faceting formula}

\item{position}{- ggplot2 string for bar position adjustments: "dodge", "stack"}

\item{xlab}{- x axis label}

\item{ylab}{- y axis label}

\item{units}{- combined with y axis label}

\item{lnscale}{- flag to plot on ln-scale}

\item{title}{- plot title}

\item{guideTitleColour}{- title for colour guide}

\item{guideTitleFill}{- title for fill guide}

\item{showPlot}{- flag to show plot immediately}
}
\value{
ggplot2 object
}
\description{
Function to plot a melted dataframe using a barchart format.
}
\details{

}

