% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDFR.Bubbles.R
\name{plotMDFR.Bubbles}
\alias{plotMDFR.Bubbles}
\title{Create or add to a circle plot of age or size comp.s.}
\usage{
plotMDFR.Bubbles(mdfr, x = NULL, y = NULL, value.var = "val",
  agg.formula = NULL, agg.function = sum, ..., colour = NULL,
  faceting = NULL, units = "", xlab = "", ylab = "", title = "",
  alpha = 0.5, maxBubbleSize = 6, useColourGradient = FALSE,
  guideTitleColour = "", showPlot = FALSE)
}
\arguments{
\item{x}{- column name for x-axis values}

\item{y}{- column name for y-axis values}

\item{value.var}{- column name for values to aggregate (value.var in cast)/plot as circles}

\item{agg.formula}{- aggregation formula (left-hand side of cast formula)}

\item{agg.function}{- aggregation function (fun.aggregate in cast)}

\item{...}{- further arguments passed to aggregating function}

\item{colour}{- column name to which colour aesthetic is mapped}

\item{faceting}{- faceting formula}

\item{units}{- units for bubble size scale}

\item{xlab}{- x axis label}

\item{ylab}{- y axis label}

\item{title}{- plot title}

\item{alpha}{- transparency level}

\item{maxBubbleSize}{- max bubble size}

\item{useColourGradient}{- flag (T/F) to use a color gradient for bubble color}

\item{guideTitleColour}{- title for colour guide}

\item{showPlot}{- flag to show plot immediately}
}
\value{
ggplot2 object
}
\description{
Creates or adds to a circle plot of age or size comp.s.
}

