% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotMeanSizeComps.r
\name{plotMeanSizeComps}
\alias{plotMeanSizeComps}
\title{Plot comparison of observed and model size comps averaged over time}
\usage{
plotMeanSizeComps(size.bins, obs.comp, prd.comp, cols = NULL, ymx = NULL,
  xlab = "year", ylab = "size mmCW", pch = 21, lty = 1, lwd = 1,
  clr = "black", CI = 0.8, bar.width = 1, addToPlot = FALSE)
}
\arguments{
\item{prd.comp}{-}

\item{cols}{-}

\item{ymx}{-}

\item{xlab}{-}

\item{ylab}{-}

\item{pch}{-}

\item{lty}{-}

\item{lwd}{-}

\item{clr}{-}

\item{CI}{-}

\item{bar.width}{-}

\item{addToPlot}{-}

\item{sizebins}{-}

\item{obs.comps}{-}
}
\description{

}

