% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelResults.ScalarParams.R
\name{plotModelResults.ScalarParams}
\alias{plotModelResults.ScalarParams}
\title{Function to plot parameter values and associated uncertainty from different TCSAM2015 models.}
\usage{
plotModelResults.ScalarParams(dfr, vfr = NULL, nc = 3, nr = 4,
  showPlot = TRUE, pdf = NULL, verbose = FALSE)
}
\arguments{
\item{dfr}{- dataframe from call to extractModelResults.Params}

\item{vfr}{- dataframe from call to extractModelResults.StdDevs}

\item{nc}{- number of columns of plots per page}

\item{nr}{- number of rows of plots per page}

\item{showPlot}{- flag to show plots}

\item{pdf}{- file name for printing plots to a pdf file (or NULL to print to screen)}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
- list with plots as elements
}
\description{
This function plots parameters values, together with their limits
(if any) and the posterior distributions implied by their estimated standard,
errors from several TCSAM2015 models.
}

