% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotMulti.GG.R
\name{plotMulti.GG}
\alias{plotMulti.GG}
\title{Plot multiple ggplot graphs on the same page.}
\usage{
plotMulti.GG(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{- multiple ggplot objects}

\item{plotlist}{- a list of ggplot objects}

\item{cols}{- number of columns in layout}

\item{layout}{- a matrix specifying the layout. If present, 'cols' is ignored.}
}
\description{
Plot multiple ggplot graphs on the same page.
}
\details{
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.

Based on the 'multiplot' function from the "Cookbook for R" (www.cookbook-r.com).
}

