% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotNatZ.r
\name{plotNatZ}
\alias{plotNatZ}
\title{Plot numbers-at-size as circles}
\usage{
plotNatZ(nAtZ)
}
\arguments{
\item{nAtZ}{- 5-d array with dim.s sex,maturity state, shell condition, year, size}
}
\description{
Plots numbers at size (5-d array with dim.s
sex,maturity state, shell condition, year, size) as circles
}
\details{

}

