% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotNM.ByPC}
\alias{plotNM.ByPC}
\title{Plot natural mortality estimates by parameter combination.}
\usage{
plotNM.ByPC(res, dims = c("PC", "SEX", "MATURITY"))
}
\arguments{
\item{res}{- model results list from TCSAM 2015 model}

\item{dims}{- dimension names for natural mortality results}
}
\description{
Function to plot natural mortality estimates by parameter combination.
}
\details{

}

