% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotObjFunValues.Data.R
\name{plotObjFunValues.Data}
\alias{plotObjFunValues.Data}
\title{Plot data values (surveys and fisheries) used in the objective function, grouped by category.}
\usage{
plotObjFunValues.Data(mdfr, variable = "objfun", ggtheme = theme_grey(),
  showPlot = FALSE)
}
\arguments{
\item{mdfr}{- melted dataframe of data values from call to getObjFunValues.getData(...)}

\item{variable}{- name of variable to plot}

\item{ggtheme}{- a ggplot2 theme}

\item{showPlot}{- flag to show plots}
}
\value{
list of ggplot2 objects corresponding to different categories of priors.
}
\description{
Function to plot data values (surveys and fisheries) used in the objective function.
}

