% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotObjFunValues.CatchData.R, R/plotObjFunValues.Priors.R
\name{plotObjFunValues.Priors}
\alias{plotObjFunValues.Priors}
\title{Plot prior values used in the objective function, grouped by category.}
\usage{
plotObjFunValues.Priors(mdfr, variable = "objfun", ggtheme = theme_grey(),
  showPlot = FALSE, verbose = FALSE)

plotObjFunValues.Priors(mdfr, variable = "objfun", ggtheme = theme_grey(),
  showPlot = FALSE, verbose = FALSE)
}
\arguments{
\item{mdfr}{- melted dataframe of prior values from call to getObjFunValues.Priors(...)}

\item{variable}{- name of variable to plot}

\item{ggtheme}{- a ggplot2 theme}

\item{showPlot}{- flag to show plots}

\item{verbose}{- flag (T/F) to print diagnostic info}

\item{showPlots}{- flag to show plots}

\item{mdfr}{- melted dataframe of prior values from call to getObjFunValues.Priors(...)}

\item{variable}{- name of variable to plot}

\item{ggtheme}{- a ggplot2 theme}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
list of ggplot2 objects corresponding to different categories of priors.

list of ggplot2 objects corresponding to different categories of priors.
}
\description{
Function to plot prior values in the objective function.

Function to plot prior values in the objective function.
}

