% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotObjFunValues.R
\name{plotObjFunValues}
\alias{plotObjFunValues}
\title{Plot all objective function values.}
\usage{
plotObjFunValues(res, variable = "objfun", ggtheme = ggplot2::theme_grey(),
  showPlots = FALSE)
}
\arguments{
\item{res}{- model results object or list of results objects}

\item{variable}{- name of variable to plot}

\item{ggtheme}{- a ggplot2 theme}

\item{showPlots}{- flag to show plots}
}
\value{
list of ggplot2 objects corresponding to penalties, priors, and data plots.
}
\description{
Function to plot all objective function values.
}
\details{

}

