% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotParameters.NumberVector}
\alias{plotParameters.NumberVector}
\title{Plot initial and final parameter values associated with a ...NumberVector info list object.}
\usage{
plotParameters.NumberVector(nv, label = "", mcmc = NULL, delta = 0.005,
  plotConsts = FALSE, xlab = "", ylab = "", pchs = 21:25,
  clrs = c("black", "blue", "green", "grey", "cyan"))
}
\arguments{
\item{label}{- the ...NumberVector variable name}

\item{mcmc}{- mcmc results corresponding to vv}

\item{plotConsts}{- flag (T/F) to plot parameters even though they were not estimated}

\item{xlab}{- the x axis label}

\item{ylab}{- the y-axis label}

\item{pchs}{- vector of point symbols to use}

\item{clrs}{- vector of colors to use}

\item{vv}{- the ...NumberVector info list object}
}
\description{
Function to plot initial and final parameter values associated with a ...NumberVector info list object.
}
\details{

}

