% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotParameters.R
\name{plotParameters}
\alias{plotParameters}
\title{Plot initial & final values, limits, and priors for model parameters.}
\usage{
plotParameters(res, res.mcmc = NULL, nv.mfrow = NULL, vv.mfrow = NULL)
}
\arguments{
\item{res}{- TCSAM2015 model results list object}

\item{res.mcmc}{- list object for compiled MCMC results}

\item{nv.mfrow}{- number of rows/page for plotting NumberVector parameters}

\item{vv.mfrow}{- number of rows/page for plotting VectorVector parameters}
}
\description{
Function to plot initial & final values, limits, and priors for model parameters.
}
\details{

}

