% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParameters.VectorVector.R
\name{plotParameters.VectorVector}
\alias{plotParameters.VectorVector}
\title{Plot initial and final parameter values associated with a ...VectorVector info list object.}
\usage{
plotParameters.VectorVector(vv, label = "", xlab = "", ylab = "",
  pchs = 21:25, clrs = c("black", "blue", "green", "grey", "cyan"),
  verbose = FALSE)
}
\arguments{
\item{vv}{- the ...VectorVector info list object}

\item{label}{- the ...VectorVector variable name}

\item{xlab}{- the x axis label}

\item{ylab}{- the y-axis label}

\item{pchs}{- vector of point symbols to use}

\item{clrs}{- vector of colors to use}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\description{
Function to plot initial and final parameter values associated with a ...VectorVector info list object.
}

