% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPDFRects.R
\name{plotPDFRects}
\alias{plotPDFRects}
\title{Plot a pdfRects object on an existing plot.}
\usage{
plotPDFRects(o, y = NULL, colorscale = "jet", alpha = 0.5, add = TRUE)
}
\arguments{
\item{o}{- the pdfRects object}

\item{colorscale}{- the name of the colorscale to use}

\item{alpha}{- the transparency level for the resulting image}
}
\description{
Function to plot an objec of class 'pdfRects' on an existing plot.
}
\details{

}

