% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotPrGrowth.ByPC}
\alias{plotPrGrowth.ByPC}
\title{Plot pr(z'|x,z) estimates by parameter combination.}
\usage{
plotPrGrowth.ByPC(res)
}
\arguments{
\item{res}{- model results list from TCSAM 2015 model}
}
\description{
Function to plot pr(z'|x,z) estimates by parameter combination.
}
\details{

}

