% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPrMaturity.byPC.R
\name{plotPrMaturity.ByPC}
\alias{plotPrMaturity.ByPC}
\title{Plot pr(mature|z) estimates by parameter combination.}
\usage{
plotPrMaturity.ByPC(res)
}
\arguments{
\item{res}{- model results list from TCSAM 2015 model}
}
\description{
Function to plot pr(mature|z) estimates by parameter combination.
}
\details{

}

