% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotRecruits.SexRatio.R
\name{plotRecruits.SexRatio}
\alias{plotRecruits.SexRatio}
\title{Plot time series of TCSAM 2015 model sex ratio for recruits}
\usage{
plotRecruits.SexRatio(res, xlab = "model year",
  ylab = "recruitment sex fraction (males)", xlim = NULL, ylim = c(0, 1),
  clr = "blue", pts.plot = TRUE, pts.type = 21, pts.siz = 8,
  line.plot = TRUE, line.wd = 2, line.type = 1, overplot = FALSE)
}
\arguments{
\item{res}{- list object from TCSM 2015 model report file}

\item{xlab}{- x axis title}

\item{ylab}{- y axis title}

\item{xlim}{- x axis limits}

\item{ylim}{- y axis limits}

\item{clr}{- line/points color}

\item{pts.plot}{- flag (T/F) to plot points}

\item{pts.type}{- type of point to plot (pch from 'par')}

\item{pts.siz}{- point size}

\item{line.plot}{- flag (T/F) to plot line}

\item{line.wd}{- line width (lwd from par)}

\item{line.type}{- type of line (lty from 'par')}

\item{overplot}{- flag (T/F) to plot the data on an existing plot}
}
\description{
Plots the recruitment sex ratio time series from a TCSAM 2015 model run.
}
\details{

}

