% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotSelFcns.ByPC}
\alias{plotSelFcns.ByPC}
\title{Plot selectivity functions by parameter combination.}
\usage{
plotSelFcns.ByPC(res, by = 5, nperpage = 2, pchs = c(21, 22, 23, 24, 25),
  clrs = c("black", "blue", "green", "cyan", "grey"))
}
\arguments{
\item{res}{- a TCSAM2015 model report object (class='tcsam2015')}

\item{by}{- number of functions to include in one page}

\item{nperpage}{- number of plots per page}

\item{pchs}{- vector of point symbol ids}

\item{clrs}{- vector of colors to use}
}
\description{
Function to plot selectivity functions by parameter combination.
}
\details{

}

