% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSelFcns.ByYear.R
\name{plotSelFcns.ByYear}
\alias{plotSelFcns.ByYear}
\title{Compare selectivity functions from TCSAM2015 and rsimTCSAM model runs by year}
\usage{
plotSelFcns.ByYear(repObj = NULL, nrow = 3, ncol = 4, showPlot = TRUE,
  pdf = NULL, width = 8, height = 6)
}
\arguments{
\item{repObj}{- single TCSAM2015 model report object, or named list of such}

\item{showPlot}{- flag to show/print plots immediately}

\item{pdf}{- name of pdf file to record plot output to}

\item{width}{- pdf page width (in inches)}

\item{height}{- pdf page width (in inches)}
}
\value{
list of ggplot2 objects
}
\description{
Function to compare selectivity functions from TCSAM2015 and rsimTCSAM model runs by year.
}

