% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotSingleSizeCompGG.R
\name{plotSingleSizeCompGG}
\alias{plotSingleSizeCompGG}
\title{Plot size comps from a single year (possibly for multiple models).}
\usage{
plotSingleSizeCompGG(n_xmsz = NULL, res = NULL, component = "pop.quants",
  year = NULL, mdl = "", title = "", showPlot = TRUE)
}
\arguments{
\item{n_xmsz}{- array (or list of arrays) dimensioned xmsz}

\item{res}{- TCSAM2015 results object (or list of such)}

\item{component}{- object name in res from which to extract n_xmsyz}

\item{year}{- year to extract size comps from component in res}

\item{mdl}{- model name (if single model is given)}

\item{title}{- title for plot}

\item{showPlot}{- flag to show plot immediately}
}
\value{
ggplot2 object
}
\description{
Function to plot size comps from a single year (possibly for multiple models).
}
\details{

}

