% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotSizeCompsComparisons}
\alias{plotSizeCompsComparisons}
\title{Plot comparisons of model and data size compositions.}
\usage{
plotSizeCompsComparisons(name, od, md, ncol = 2, showPlots = TRUE)
}
\description{
Function to plot comparisons of model and data size compositions.
}
\details{

}

