% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSizeCompsComparisons0.R
\name{plotSizeCompsComparisons0}
\alias{plotSizeCompsComparisons0}
\title{Plot comparisons of model and data size compositions.}
\usage{
plotSizeCompsComparisons0(name, od, md, normalize = TRUE, label = "",
  ncol = 2, showPlots = TRUE)
}
\arguments{
\item{name}{- name of size comps source}

\item{od}{- observed size comps list object (may be NULL)}

\item{md}{- model (predicted) size comps list object (may be NULL)}

\item{label}{- label for title}

\item{ncol}{- number of columns per page for plots}

\item{showPlots}{- flag (T/F) to show plots}
}
\value{
list of ggplot objects comprising the graph pages to plot
}
\description{
Function to plot comparisons of model and data size compositions.
}

