% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSizeCompsComparisons2.R
\name{plotSizeCompsComparisons2}
\alias{plotSizeCompsComparisons2}
\title{Plot comparisons of model and data size compositions by year and sex.}
\usage{
plotSizeCompsComparisons2(name, od, md, label = "", disAggBy = c("sx", "ms",
  "sc"), normalize = TRUE, normBy = NULL, ggtheme = theme_grey(),
  ncol = 2, nrow = 5, showPlot = FALSE)
}
\arguments{
\item{name}{- name of size comps source}

\item{od}{- observed size comps list object (may be NULL)}

\item{md}{- model (predicted) size comps list object (may be NULL)}

\item{label}{- label for title}

\item{disAggBy}{- vector of categories to keep disaggregated ('sx','ms',and/or 'sc') or NULL to aggregate over all}

\item{normalize}{- flag (T/F) to normalize size comps}

\item{normBy}{- vector of categories to normalize by ('sx','ms',and/or 'sc') or NULL to normalize over all}

\item{ggtheme}{- ggplot2 theme to use}

\item{ncol}{- number of columns per page for plots}

\item{nrow}{- number of rows per page for plots}

\item{showPlot}{- flag (T/F) to show plots immediately}
}
\value{
list of ggplot objects comprising the graph pages to plot
}
\description{
Function to plot comparisons of model and data size compositions by year for
specified sexes. Maturity/shell condition combinations
for a given year x sex combination are distinguished by different colors.
}

