% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotSpB.R
\name{plotSpB}
\alias{plotSpB}
\title{Plot time series of TCSAM 2014 model spawning biomass}
\usage{
plotSpB(res, logscale = FALSE, xlab = "model year",
  ylab = "spawning biomass (1000's t)", xlim = NULL, ylim = NULL,
  clr = c("red", "blue"), pts.plot = TRUE, pts.type = 21, pts.siz = 8,
  line.plot = TRUE, line.wd = 2, line.type = c(2, 1), overplot = FALSE)
}
\arguments{
\item{res}{- results object from TCSM 2014 model output}

\item{logscale}{- flag (T/F) to plot on ln-scale}

\item{xlab}{- x axis title}

\item{ylab}{- y axis title}

\item{xlim}{- x axis limits}

\item{ylim}{- y axis limits}

\item{clr}{- line/points color}

\item{pts.plot}{- flag (T/F) to plot points}

\item{pts.type}{- type of point to plot (pch from 'par')}

\item{pts.siz}{- point size}

\item{line.plot}{- flag (T/F) to plot line}

\item{line.wd}{- line width (lwd from par)}

\item{line.type}{- type of line (lty from 'par')}

\item{overplot}{- flag (T/F) to plot the data on an existing plot}
}
\description{
Plots the spawning biomass time series from a TCSAM 2014 model run.
}
\details{


Plots spawning biomass by sex across years in the model run. Characteristics
for points and lines (clr, type, and siz/wd) can be vectors.
}

