% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXY.R
\name{plotXY}
\alias{plotXY}
\title{Convenience function to make an x-y plot}
\usage{
plotXY(x, y, xlab = NULL, ylab = NULL, xlim = NULL, ylim = NULL,
  clr = "blue", pts.plot = TRUE, pts.type = 21, pts.siz = 8,
  line.plot = TRUE, line.wd = 2, line.type = 1, overplot = FALSE)
}
\arguments{
\item{x}{- vector of x coordinates}

\item{y}{- vector of y coordinates}

\item{xlab}{- x axis title}

\item{ylab}{- y axis title}

\item{xlim}{- x axis limits}

\item{ylim}{- y axis limits}

\item{clr}{- line/points color}

\item{pts.plot}{- flag (T/F) to plot points}

\item{pts.type}{- type of point to plot (pch from 'par')}

\item{pts.siz}{- point size}

\item{line.plot}{- flag (T/F) to plot line}

\item{line.wd}{- line width (lwd from par)}

\item{line.type}{- type of line (lty from 'par')}

\item{overplot}{- flag (T/F) to plot the data on an existing plot}
}
\description{
A convenience function to make x-y plots
}

