% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZScores.R
\name{plotZScores}
\alias{plotZScores}
\title{Plot z-scores from fits to data}
\usage{
plotZScores(afits, xlim = NULL, ylim = NULL, xlab = "years",
  ylab = "z-scores", label = "")
}
\arguments{
\item{afits}{-}

\item{xlim}{- x axis limits (computed internally if NULL)}

\item{ylim}{- y axis limits (computed internally if NULL)}

\item{xlab}{- x axis title}

\item{ylab}{- y axis title}

\item{label}{- plot label}
}
\details{
Standard graphics approach to plotting z-scores.
}

