% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZScoresForAll.R
\name{plotZScoresForAll}
\alias{plotZScoresForAll}
\title{Plot z-scores for all components}
\usage{
plotZScoresForAll(repObj, ggtheme = theme_grey(), showPlot = TRUE,
  verbose = FALSE)
}
\arguments{
\item{repObj}{- model report list object}

\item{ggtheme}{- theme for ggplot2}

\item{showPlot}{- flag to show plots immediately}

\item{verbose}{- flag (T/F) to print dagnostic info}
}
\value{
list of lists of ggplot objects
}
\description{
Function to plot z-scores for all components.
}
\details{
Requires reshape2 and ggplot2 packages.
}

