% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZScoresForCatchData.R
\name{plotZScoresForCatchData}
\alias{plotZScoresForCatchData}
\title{Plot fits to catch data (abundance, biomass, size freq.s) as z-scores.}
\usage{
plotZScoresForCatchData(fit, mc, fleet, type = "total catch",
  showPlot = TRUE, verbose = FALSE)
}
\arguments{
\item{fit}{- fits list for given catch type}

\item{mc}{- model configuration list from tcsam2015 report object (rep$mc, if rep is the report object)}

\item{fleet}{- fleet label}

\item{type}{- catch type label corresponding to the 'fit' object}

\item{showPlot}{- flag (T/F) to show plots immediately}

\item{verbose}{- flag (T/F) to print dagnostic info}
}
\value{
list of ggplot2 objects
}
\description{
Function to plot fits to catch data (abundance, biomass, size freq.s) as z-scores.
}

