% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotZScoresForFisheries.R
\name{plotZScoresForFisheries}
\alias{plotZScoresForFisheries}
\title{Plot model fits to abundance, biomass and size frequencies as z-scores for fishery data components.}
\usage{
plotZScoresForFisheries(res, showPlot = FALSE)
}
\arguments{
\item{res}{- model results list object}
}
\description{

}

