% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZScoresForFleets.R
\name{plotZScoresForFleets}
\alias{plotZScoresForFleets}
\title{Plot model fits to abundance, biomass and size frequencies as z-scores for fishery data components.}
\usage{
plotZScoresForFleets(repObj, type = "fishery", showPlot = TRUE,
  verbose = FALSE)
}
\arguments{
\item{repObj}{- model report list object}

\item{type}{- fleet type ('fishery' or 'survey')}

\item{showPlot}{- flag (T/F) to show plots immediately}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
list by fishery of lists with ggplot objects
}

