% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotZScoresForSurveys}
\alias{plotZScoresForSurveys}
\title{Plot model fits to abundance, biomass and size frequencies as z-scores for surveys.}
\usage{
plotZScoresForSurveys(res)
}
\arguments{
\item{res}{- model results list object}
}
\description{

}

