% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotZScoresGG.R
\name{plotZScoresGG}
\alias{plotZScoresGG}
\title{Plot z-scores from fits to data}
\usage{
plotZScoresGG(afits, ci = 0.95, xlim = NULL, ylim = NULL,
  xlab = "years", ylab = "z-scores", label = "", ggtheme = theme_grey(),
  showPlot = FALSE)
}
\arguments{
\item{afits}{-}

\item{ci}{- confidence interval to plot}

\item{xlim}{- x axis limits (computed internally if NULL)}

\item{ylim}{- y axis limits (computed internally if NULL)}

\item{xlab}{- x axis title}

\item{ylab}{- y axis title}

\item{label}{- title for graph page}

\item{ggtheme}{- theme for ggplot2}

\item{showPlot}{- flag to print plot immediately}
}
\value{
list of ggplot2 objects
}
\description{
Function to plot z-scores from fits to data using ggplot2.
}
\details{

}

