% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readParamsCSV.R
\name{readParamsCSV}
\alias{readParamsCSV}
\title{Read a TCSAM2015 model parameters csv file and return a dataframe.}
\usage{
readParamsCSV(csvFile = NULL)
}
\arguments{
\item{csvFile}{- parameters csv file from a TCSAM2015 model run. can be NULL.}
}
\value{
a dataframe (or NULL).
}
\description{
Function to read a TCSAM2015 parameters csv file and return a dataframe.
}
\details{
If csvFile is NULL, the user will be prompted to identify a 
TCSAM2015 model parameters csv file to read. Uses functions
\itemize{
 \item \code{wtsUtilities::selectFile(...)}
}
}

