% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readParFile.R
\name{readParFile}
\alias{readParFile}
\title{Create a TCSAM2015 par object from a par file.}
\usage{
readParFile(in.par = NULL)
}
\arguments{
\item{in.par}{= filename of par file}
}
\value{
a prs object (a dtaframe of class 'tcsam2015.par') corresponding to the par file (or NULL if file does not exist)
}
\description{
Function to create a TCSAM2015 par object from a par file.
}
\details{
Uses \code{wtsUtilities::selectFile} to open a file dialog if in.par is NULL.
}

