% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readReportFile.R
\name{readReportFile}
\alias{readReportFile}
\title{Read a TCSAM2015 model report (.rep) file.}
\usage{
readReportFile(repfile = NULL)
}
\arguments{
\item{repfile}{- report file from a TCSAM2015 model run to source. can be NULL.}
}
\value{
report model object (a list). The returned object will be a list of class 'tcsam2015'.
}
\description{
Function to read a TCSAM2015 report file (.rep).
}
\details{
If repfile is NULL, the user will be prompted to identify a 
TCSAM2015 model report file from which to source the results object.
}

