% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatTimeBlocks.R
\name{reformatTimeBlocks}
\alias{reformatTimeBlocks}
\title{Reformat rsim, tcsam notation for time blocks to a standard format.}
\usage{
reformatTimeBlocks(tb0, dims)
}
\arguments{
\item{tb0}{- time block string to reformat}

\item{dims}{- model 'dims' list from tscam or rsim object}
}
\value{
time block string in "standardized" format
}
\description{
eformat rsim, tcsam notation for time blocks to a standard format.
}

