% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runProfiles.R
\name{runProfiles}
\alias{runProfiles}
\title{Function to run profiles on a TCSAM2015 model.}
\usage{
runProfiles(os = "osx", path = ".", model = "tcsam2015",
  path2model = "", configFileTemplate = NULL, controlFileTemplate = NULL,
  profiles = NULL, runSeqs = TRUE, numRuns = 4, mcmc = FALSE,
  mc.N = 1e+06, mc.save = 1000, mc.scale = 1000, plotResults = TRUE)
}
\arguments{
\item{os}{- 'win' or 'mac' or 'osx'}

\item{path}{- path to parent folder for model output}

\item{model}{- TCSAM2015 model executable name}

\item{path2model}{- path to model executable}

\item{profiles}{- dataframe with variables to profile on in each row}

\item{runSeqs}{- TRUE to run sequences, FALSE to jitter}

\item{numRuns}{- number of sequences or jitters to run}

\item{mcmc}{- flag (T/F) to run mcmc on "best" model}

\item{mc.N}{- number of mcmc iterations to make}

\item{mc.save}{- number of iterations to skip when saving mcmc calculations}

\item{mc.scale}{- number of iterations to adjust scale for mcmc calculations}

\item{plotResults}{- T/F to plot final results within each profile using \code{???}}

\item{configFile}{- path to model configuration file template}

\item{controlFile}{- path to model control file template}
}
\value{
- par file dataframe
}
\description{
This functions runs profiles on a TCSAM2015 model.
}
\details{
This function creates a shell script ('./tmp.sh') in the
working directory and uses it to run a version of the TCSAM2015 model.
Initial model parameters can be jittered based on the system clock time as a seed
to the random number generator. The seed and final objective function value are
saved for each model run in a csv file (the value of out.csv).
}

