% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTCSAM2015.R
\name{runTCSAM2015}
\alias{runTCSAM2015}
\title{Function to run TCSAM2015.}
\usage{
runTCSAM2015(os = "osx", path = ".", model = "tcsam2015",
  path2model = "", configFile = "", pin = FALSE, hess = FALSE,
  mcmc = FALSE, mc.N = 1e+06, mc.save = 1000, mc.scale = 1000,
  jitter = FALSE, jit.seed = NULL, calcOFL = FALSE, plotResults = hess,
  cleanup = TRUE)
}
\arguments{
\item{os}{- 'win' or 'mac' or 'osx'}

\item{path}{- path for model output}

\item{model}{- TCSAM2013 model executable name}

\item{path2model}{- path to model executable}

\item{configFile}{- filename (including path) to model configuration file}

\item{pin}{- T/F to use a pin file}

\item{hess}{- T/F to compute hessian (and .std file)}

\item{mcmc}{- T/F to run mcmc}

\item{mc.N}{- number of mcmc iterations to do}

\item{mc.save}{- number of iterations to skip when saving mcmc calculations}

\item{mc.scale}{- number of iterations to adjust scale for mcmc calculations}

\item{jitter}{- T/F to jitter parameters}

\item{jit.seed}{- seed for random number generator (or NULL)}

\item{calcOFL}{- flag (T/F) to perform OFL calculations}

\item{plotResults}{- T/F to plot results using \code{plotTCSAM2015I(...)}}

\item{cleanup}{- flag (T/F) to clean up unnecessary files}
}
\value{
- dataframe of class 'tcam2015.par', with 2 columns (name, value) with jitter jit.seed (if jittered) 
and par file info, or NULL if par file does not exist.
}
\description{
This function runs a TCSAM2015 model once.
}
\details{
This function creates a shell script ('./tmp.sh') in the
working directory and uses it to run a version of the TCSAM2015 model.\cr
Initial model parameters can be jittered based on the system clock time as a jit.seed
to the random number generator. The jit.seed and final objective function value are
saved for each model run in a csv file (the value of out.csv).


If the path associated with \code{configFile} is a relative one, it should
be relative to the \code{path} variable. If showPlot=TRUE, the report file, prs file, 
and std files are read in and the associated objects are saved to 'ModelResults.RData'
as repObj, prsObj, and stdObj, respectively.
}

