% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selectReportFile.R
\name{selectReportFile}
\alias{selectReportFile}
\title{Get a TCSAM2015 model report file}
\usage{
selectReportFile()
}
\value{
path to selected TCSAM2015 model report file, or NULL (if user canceled selection).
}
\description{
Allows the user to select a TCSAM2015 model report file using a file dialog.
}
\details{

}

