% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sumByDims}
\alias{sumByDims}
\title{Sum array by specified dimensions}
\usage{
sumByDims(n, dims = 1)
}
\arguments{
\item{n}{- array to sum}

\item{dims}{- vector of indices for dimension to sum by}
}
\value{
array with dimensions used for summing
}
\description{

}

