% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeModelResultsToCSV.ObjFunValues.R
\name{writeModelResultsToCSV.ObjFunValues}
\alias{writeModelResultsToCSV.ObjFunValues}
\title{Write all objective function values to csvs}
\usage{
writeModelResultsToCSV.ObjFunValues(repObjs,
  csv = "ModelComparisons.ObjFunValues", verbose = FALSE)
}
\arguments{
\item{repObjs}{- model report object or list of report objects}

\item{csv}{- base filename for csvs to export dataframes to}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
list w two dataframes with objective function components, 
one for penalties and priors (...PP.csv) and one for data (...Data.csv).
}
\description{
Function to write all objective function values to csvs.
}

