% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeModelResultsToCSV.ParamEsts.R
\name{writeModelResultsToCSV.ParamEsts}
\alias{writeModelResultsToCSV.ParamEsts}
\title{Function to write parameter values and standard deviations from different TCSAM2015 models to a csv file.}
\usage{
writeModelResultsToCSV.ParamEsts(tcsams, dp = 0.01,
  csv = "ModelComparisons.Params.csv", verbose = FALSE)
}
\arguments{
\item{tcsams}{- list of TCSAM2015 model results objects (each is a list with elements 'prsObj' and 'stdObj')}

\item{dp}{- percent difference between parameter value and upper/lower limits used to flag outliers}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
- dataframe returned by \code{extractModelResults.ParamsPlusStdDevs}
}
\description{
This function extracts parameters values, together with their limits
(if any) and their associated standard deviations from several TCSAM2015 models as a dataframe
and writes it to a csv file.
}

