% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeModelResultsToCSV.PopProcesses.R
\name{writeModelResultsToCSV.PopProcesses}
\alias{writeModelResultsToCSV.PopProcesses}
\title{Write population processes from model results from TCSAM2015 and rsimTCSAM model runs to a csv file}
\usage{
writeModelResultsToCSV.PopProcesses(tcsams = NULL, rsims = NULL,
  csv = "ModelResults.PopProcesses", verbose = FALSE)
}
\arguments{
\item{tcsams}{- single TCSAM2015 model report object, or named list of such}

\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{csv}{- base name for csv files to write to}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
nothing
}
\description{
Function to write population processes from model results from TCSAM2015 and rsimTCSAM model runs to a csv file.
}
\details{
Extracts natural mortality, mean growth increments, molt-to-maturity, and recruitment 
size distribution information.
}

