% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeModelResultsToCSV.Surveys.R
\name{writeModelResultsToCSV.Surveys}
\alias{writeModelResultsToCSV.Surveys}
\title{Write survey quantities from model results from TCSAM2015 and rsimTCSAM model runs to a csv file.}
\usage{
writeModelResultsToCSV.Surveys(tcsams = NULL, rsims = NULL,
  cast.formula = "v+y+x+m", csv = "ModelResults.Surveys", verbose = FALSE)
}
\arguments{
\item{tcsams}{- single TCSAM2015 model report object, or named list of such}

\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{cast.formula}{- formula for casting (aggregating)}

\item{csv}{- base name for csv files to write to}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
nothing
}
\description{
Function to write survey quantities from model results from TCSAM2015 and rsimTCSAM model runs to a csv file.
}
\details{
none.
}

