% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeModelResultsToCSV.SurveySelFcns.R
\name{writeModelResultsToCSV.SurveySelFcns}
\alias{writeModelResultsToCSV.SurveySelFcns}
\title{Write survey selectivity functions to csv from TCSAM2015 and rsimTCSAM model runs.}
\usage{
writeModelResultsToCSV.SurveySelFcns(tcsams = NULL, rsims = NULL,
  years = list(`NMFS Trawl Survey` = 2014), cast.factors = "x",
  csv = "ModelResults.SurveySelFcns", verbose = FALSE)
}
\arguments{
\item{tcsams}{- single TCSAM2015 model report object, or named list of such}

\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{years}{- list, by fishery name, of years to plot}

\item{cast.factors}{- x,m,s factors to include in output file. Non-specified factor levels will be averaged over.}

\item{csv}{- base name for csv files to write to}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
list of ggplot2 objects
}
\description{
Function to write survey selectivity functions to csv from TCSAM2015 and rsimTCSAM model runs.
}
\details{
Factors listed in the cast.factors formula will be included as levels in output file.
}

