% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compareModels.R
\name{compareModels}
\alias{compareModels}
\title{Compare TCSAM2015 and rsimTCAM model runs.}
\usage{
compareModels(tcsam = NULL, rsim = NULL, pdf = "compareModels.pdf",
  width = 8, height = 6)
}
\arguments{
\item{tcsam}{- single TCSAM2015 model results object, or named list of such}

\item{rsim}{- single rsimTCSAM results object, or named list of such}

\item{pdf}{- name of pdf file to record plot output to}

\item{width}{- pdf page width (in inches)}

\item{height}{- pdf page width (in inches)}
}
\value{
list of ggplot2 objects
}
\description{
Function to compare TCSAM2015 and rsimTCAM model runs.
}
\details{

}

