% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.ParamEsts.R
\name{extractModelResults.Params}
\alias{extractModelResults.Params}
\title{Function to extract parameter values from different models.}
\usage{
extractModelResults.Params(objs, dp = 0.01)
}
\arguments{
\item{objs}{- list of model results objects (each is a list with elements 'prsObj' and 'stdObj')}

\item{dp}{- percent difference between parameter value and upper/lower limits used to flag outliers}
}
\value{
- dataframe with the parameter information
}
\description{
This function extracts parameters values, together with their limits
(if any) from several models.
}

