% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMDFR.Special.R
\name{getMDFR.NatMort}
\alias{getMDFR.NatMort}
\title{Get natural mortality rates from model results from TCSAM2015 and rsimTCSAM model runs as a dataframe}
\usage{
getMDFR.NatMort(tcsams, rsims, verbose = FALSE)
}
\arguments{
\item{tcsams}{- single TCSAM2015 model report object, or named list of such}

\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
nothing
}
\description{
Function to get natural mortlaity rates from model results from TCSAM2015 and rsimTCSAM model runs as a dataframe.
}
\details{
Extracts natural mortality.
}

