% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrs.R
\name{getPrs}
\alias{getPrs}
\title{Get a TCSAM2015 prs object by reading initial and final parameters csv files.}
\usage{
getPrs(type = "all", inp.dir = ".")
}
\arguments{
\item{type}{- 'all' or 'active'}
}
\value{
A prs model object (a dataframe). The returned object will be a list of class 'tcsam2015.prs'.
}
\description{
Function to get a TCSAM2015 prs object by reading initial and final parameters csv file.
}
\details{
To create the prs object, this function reads 2 csv-type parameter value files,
one associated with the initial parameter values and one associated with the final values.
The user can select to return either all the parameters, or only the active parameters (ones
with phase > 0). The returned object will be a dataframe of class 'tcsam2015.prs'.
}

