% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotCatchData}
\alias{plotCatchData}
\title{Plot catch data}
\usage{
plotCatchData(name = NULL, obs = NULL, mod = NULL, logscale = FALSE,
  ylab = "")
}
\arguments{
\item{name}{- fishery or survey name}

\item{obs}{- list of observed data}

\item{mod}{- list of model-predicted data}

\item{logscale}{- flag (T/F) to plot on ln-scale}

\item{ylab}{- base title for y axis}
}
\description{
Plot catch data (aggregate catch, catch-at-size) from fisheries or surveys
}
\details{

}

