% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCatchData.R
\name{plotCatchData}
\alias{plotCatchData}
\title{Plot catch data}
\usage{
plotCatchData(name = NULL, obs = NULL, mod = NULL, logscale = FALSE,
  normalize = TRUE, label = "", showPlot = FALSE)
}
\arguments{
\item{name}{- fishery or survey name}

\item{obs}{- list of observed data}

\item{mod}{- list of model-predicted data}

\item{logscale}{- flag (T/F) to plot on ln-scale}

\item{label}{- base label for plots}

\item{showPlot}{- flag to show plots immediately}
}
\value{
multi-level list of ggplot2 objects
}
\description{
Plot catch data (aggregate catch, catch-at-size) from fisheries or surveys
}
\details{

}

