% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotDataModelComparisons.R
\name{plotDataModelComparisons}
\alias{plotDataModelComparisons}
\title{Plot model-simulated and observed data for comparison.}
\usage{
plotDataModelComparisons(res)
}
\arguments{
\item{res}{- results list object from report file}
}
\description{

}

