% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotDataModelComparisons.R
\name{plotDataModelComparisons}
\alias{plotDataModelComparisons}
\title{Plot model-estimated and observed data for comparison.}
\usage{
plotDataModelComparisons(repObj, showPlot = FALSE)
}
\arguments{
\item{repObj}{- results list object from report file}

\item{showPlot}{- flag to show plot immediately}
}
\value{
multi-level list with ggplot2 objects
}
\description{

}

