% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotFisheriesResults.R
\name{plotFisheriesResults}
\alias{plotFisheriesResults}
\title{Plot model results for fisheries.}
\usage{
plotFisheriesResults(res, showPlot = FALSE)
}
\arguments{
\item{res}{- results list from a TCSAM2015 model run}

\item{showPlot}{- flag (T/F) to show plots}
}
\value{
nested list of ggplot objects
}
\description{
Function to plot model results for fisheries.
}
\details{

}

