% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFisheryCatchesGG.R
\name{plotFisheryCatchesGG}
\alias{plotFisheryCatchesGG}
\title{Plot time series of the fishery catch (numbers: captured, discards, retained) by fishery and sex.}
\usage{
plotFisheryCatchesGG(res, showPlot = TRUE)
}
\arguments{
\item{res}{- results list from a TCSAM2015 model run}

\item{showPlot}{- flag (T/F) to show plots}
}
\value{
list of ggplot objects
}
\description{
Plot time series of the fishery catch (numbers: captured, discards, retained) by fishery and sex.
}

