% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotFishingRatesGG.R
\name{plotFishingRatesGG}
\alias{plotFishingRatesGG}
\title{Plot time series of the fully-selected fishery capture rates.}
\usage{
plotFishingRatesGG(res, ggtheme = theme_bw(), showPlot = FALSE)
}
\arguments{
\item{res}{- results list from a TCSAM2015 model run}

\item{ggtheme}{- a ggplot2 theme}

\item{showPlot}{- flag (T/F) to show plots}
}
\value{
list of ggplot objects
}
\description{
Plot time series of the fully-selected fishery capture rates.
}
\details{

}

