% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInputDataTimeframes.R
\name{plotInputDataTimeframes}
\alias{plotInputDataTimeframes}
\title{Plot an input data timeframe dataframe}
\usage{
plotInputDataTimeframes(dfr, showPlot = FALSE)
}
\arguments{
\item{dfr}{- dataframe with input data time frames}

\item{showPlot}{- flag (T/F) to print plot}
}
\value{
returns a ggplot2 object
}
\description{
A function to plot an input data timeframe dataframe.
}

