% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeanSizeCompsGG.R
\name{plotMeanSizeCompsGG}
\alias{plotMeanSizeCompsGG}
\title{Plot comparisons of mean model and data size compositions by year and sx.}
\usage{
plotMeanSizeCompsGG(name, od, md, label = "", disAggBy = c("sx", "ms",
  "sc"), normalize = TRUE, normBy = NULL, ggtheme = theme_grey(),
  showPlot = FALSE)
}
\arguments{
\item{name}{- name of size comps source}

\item{od}{- observed size comps list object (may be NULL)}

\item{md}{- model (predicted) size comps list object (may be NULL)}

\item{label}{- label for title}

\item{disAggBy}{- vector of categories to keep disaggregated ('sx','ms',and/or 'sc') or NULL to aggregate over all}

\item{normalize}{- flag (T/F) to normalize size comps}

\item{normBy}{- vector of categories to normalize by ('sx','ms',and/or 'sc') or NULL to normalize over all}

\item{ggtheme}{- ggplot2 theme to use}

\item{showPlot}{- flag (T/F) to show plots immediately}
}
\value{
list of ggplot objects comprising the graph pages to plot
}
\description{
Function to plot comparisons of mean model and data size compositions by year for
specified sexes. Maturity/shell condition combinations
for a given year x sx combination are distinguished by different colors.
}

