% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelResults.R
\name{plotModelResults}
\alias{plotModelResults}
\title{Plot TCSAM2015 model output.}
\usage{
plotModelResults(repObj = NULL, prsObj = NULL, stdObj = NULL,
  objList = NULL, ggtheme = theme_grey(), showPlot = TRUE, pdf = NULL,
  width = 14, height = 8, verbose = FALSE)
}
\arguments{
\item{repObj}{- list object based on sourcing a TCSAM2015 model report file. can be NULL or a filename.}

\item{prsObj}{- list object based on reading a TCSAM2015 active parameters csv file. can be NULL or a filename.}

\item{stdObj}{- list object based on reading a TCSAM2015 std file. can be NULL or a filename.}

\item{objList}{- list with optional elements repObj, prsObj, stdObj (an optional way to provide the Obj's)}

\item{ggtheme}{- a ggplot2 theme to use with ggplot2 plots}

\item{showPlot}{- flag to show plots immediately}

\item{pdf}{- filename for pdf output (optional)}

\item{width}{- pdf page width (in inches)}

\item{height}{- pdf page width (in inches)}

\item{verbose}{- flag (T/F) to print dagnostic info}
}
\value{
list with possible elements:\cr
* repObj - TCSAM2015 report list object (if report file was read)
* prsObj - TCSAM2015 parameters dataframe object (if parameters csv file was read)
* stdObj - TCSAM2015 std devs dataframe object (if std file was read)
* plots: multi-level list of ggplot2 objects
}
\description{
Function to plot data and results from a TCSAM2015 model run.
}
\details{
If repObj is a character string (the path to a report file), the file will be read in.
If repObj is NULL, the user will be prompted to identify a 
model report file from which to source the report object. Similarly for prsObj and stdObj.
Any of these objects obtained by reading a file will be an element in the returned list.
}

