% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotModelResults}
\alias{plotModelResults}
\title{Plot TCSAM 2015 model output.}
\usage{
plotModelResults(res = NULL, resfile = NULL, out.pdf = NULL)
}
\arguments{
\item{res}{- list object based on sourcing a TCSAM2015 model report file. can be NULL.}

\item{resfile}{- report file from a TCSAM2015 model run to source. can be NULL.}

\item{out.pdf}{- file for pdf output (optional)}
}
\value{
results model object.
}
\description{
Function to plot data and results from a TCSAM2015 model run.
}
\details{


If res is NULL, resfile will be sourced locally to provide res. If
resfile is also NULL, the user will be prompted to identify a
model report file from which to source the results object. \cr
}

