% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotObjFunValues.Penalties.R
\name{plotObjFunValues.Penalties}
\alias{plotObjFunValues.Penalties}
\title{Plot penalties used in the objective function, grouped by category.}
\usage{
plotObjFunValues.Penalties(mdfr, variable = "objfun",
  ggtheme = theme_grey(), showPlot = FALSE, verbose = FALSE)
}
\arguments{
\item{mdfr}{- melted dataframe of penalties from call to getObjFunValues.Penalties(...)}

\item{variable}{- name of variable to plot}

\item{ggtheme}{- a ggplot2 theme}

\item{showPlot}{- flag to show plots immediately}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
ggplot2 object
}
\description{
Function to plot penalties in the objective function.
}

