% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotObjFunValues.Penalties.R
\name{plotObjFunValues.Penalties}
\alias{plotObjFunValues.Penalties}
\title{Plot penalties used in the objective function, grouped by category.}
\usage{
plotObjFunValues.Penalties(mdfr, variable = "objfun",
  ggtheme = theme_grey(), showPlots = FALSE)
}
\arguments{
\item{mdfr}{- melted dataframe of penalties from call to getObjFunValues.Penalties(...)}

\item{variable}{- name of variable to plot}

\item{ggtheme}{- a ggplot2 theme}

\item{showPlots}{- flag to show plots}
}
\value{
ggplot2 object
}
\description{
Function to plot penalties in the objective function.
}
\details{

}

