% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotObjFunValues.R
\name{plotObjFunValues}
\alias{plotObjFunValues}
\title{Plot all objective function values.}
\usage{
plotObjFunValues(res, variable = "objfun", ggtheme = theme_grey(),
  showPlot = FALSE)
}
\arguments{
\item{res}{- model results object or list of results objects}

\item{variable}{- name of variable to plot}

\item{ggtheme}{- a ggplot2 theme}

\item{showPlot}{- flag to show plots immediately}
}
\value{
list of ggplot2 objects corresponding to penalties, priors, and data plots.
}
\description{
Function to plot all objective function values.
}

