% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotPopComponents.Abundance.R
\name{plotPopComponents.Abundance}
\alias{plotPopComponents.Abundance}
\title{Plot population abundance components.}
\usage{
plotPopComponents.Abundance(res, z.legal = 138, clrs = c("black", "green",
  "blue", "grey", "cyan"))
}
\arguments{
\item{res}{- model results list object}

\item{z.legal}{- legal size (mm CW)}

\item{clrs}{- color scheme for plotting}
}
\description{

}

