% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotPopQuants.R
\name{plotPopQuants}
\alias{plotPopQuants}
\title{Plot model-estimated population quantities such as natural mortality rates, maturity schedules, etc.}
\usage{
plotPopQuants(repObj, showPlot = TRUE)
}
\arguments{
\item{repObj}{- model report list from TCSAM 2015 model}

\item{showPlot}{- flag to show plots immediately}
}
\value{
list of list of ggplot2 objects
}
\description{
Function to plot model-estimated population quantities such as
natural mortality rates, maturity schedules, etc.
}
\details{

}

