% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotPrGrowth.byPC.R
\name{plotPrGrowth.ByPC}
\alias{plotPrGrowth.ByPC}
\title{Plot pr(z'|x,z) estimates by parameter combination.}
\usage{
plotPrGrowth.ByPC(res)
}
\arguments{
\item{res}{- model results list from TCSAM 2015 model}
}
\description{
Function to plot pr(z'|x,z) estimates by parameter combination.
}
\details{

}

