% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotSizeCompsComparisons1}
\alias{plotSizeCompsComparisons1}
\title{Plot comparisons of model and data size compositions by year.}
\usage{
plotSizeCompsComparisons1(name, od, md, label = "", normalize = TRUE,
  ncol = 2, nrow = 5, showPlots = TRUE)
}
\arguments{
\item{name}{- name of size comps source}

\item{od}{- observed size comps list object (may be NULL)}

\item{md}{- model (predicted) size comps list object (may be NULL)}

\item{label}{- label for title}

\item{ncol}{- number of columns per page for plots}

\item{nrow}{- number of rows per page for plots}

\item{showPlots}{- flag (T/F) to show plots}
}
\value{
list of ggplot objects comprising the graph pages to plot
}
\description{
Function to plot comparisons of model and data size compositions by year. Males are plotted
as positive frequencies, females as negative frequencies, and maturity/shell condition combinations
distinguished by different colors.
}
\details{

}

