% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotZScoresForAll.R
\name{plotZScoresForAll}
\alias{plotZScoresForAll}
\title{Plot model fits to data components}
\usage{
plotZScoresForAll(repObj, showPlot = FALSE)
}
\arguments{
\item{repObj}{- model report list object}

\item{showPlot}{- flag to show plots immediately}
}
\value{
list of lists of ggplot objects
}
\description{

}

