% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotZScoresForAll.R
\name{plotZScoresForAll}
\alias{plotZScoresForAll}
\title{Plot model fits to data components}
\usage{
plotZScoresForAll(res)
}
\arguments{
\item{res}{- model results list object}
}
\description{

}

