% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZScoresForAll.R
\name{plotZScoresForAll}
\alias{plotZScoresForAll}
\title{Plot model fits to data components}
\usage{
plotZScoresForAll(repObj, ggtheme = theme_grey(), showPlot = FALSE)
}
\arguments{
\item{repObj}{- model report list object}

\item{ggtheme}{- theme for ggplot2}

\item{showPlot}{- flag to show plots immediately}
}
\value{
list of lists of ggplot objects
}

