% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotZScoresForFisheries}
\alias{plotZScoresForFisheries}
\title{Plot model fits to abundance, biomass and size frequencies as z-scores for fishery data components.}
\usage{
plotZScoresForFisheries(res)
}
\arguments{
\item{res}{- model results list object}
}
\description{

}

