% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotZScoresForSurveys.R
\name{plotZScoresForSurveys}
\alias{plotZScoresForSurveys}
\title{Plot model fits to abundance, biomass and size frequencies as z-scores for surveys.}
\usage{
plotZScoresForSurveys(repObj, showPlot = FALSE)
}
\arguments{
\item{repObj}{- model report list object}
}
\value{
list by survey of lists with ggplot objects
}
\description{

}

