% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotZScoresGG.SizeFreqs.R
\name{plotZScoresGG.SizeFreqs}
\alias{plotZScoresGG.SizeFreqs}
\title{Plot residuals from size frequency fits}
\usage{
plotZScoresGG.SizeFreqs(fits, mc, sxs = c(mc$SXs, "ALL_SEX"),
  mss = c(mc$MSs, "ALL_MATURITY"), scs = c(mc$SCs, "ALL_SHELL"),
  label = "", ggtheme = theme_grey(), showPlot = TRUE)
}
\arguments{
\item{fits}{- list of fits}

\item{mc}{- model configuration list}

\item{sxs}{- character vector of sexes to plot}

\item{mss}{- character vector of maturity states to plot}

\item{scs}{- character vector of shell conditions to plot}

\item{label}{- plot label}

\item{ggtheme}{- ggplot2 theme}

\item{showPlot}{- flag to show (print) plot immediately on current graphics device}
}
\value{
list of list of ggplot2 plot objects
}
\description{
Plot Pearson's residuals or negative log-likelihood components
from size frequency fits using ggplot2.
}
\details{

}

