% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStdFile.R
\name{readStdFile}
\alias{readStdFile}
\title{Read a TCSAM2015 model std file and return a dataframe.}
\usage{
readStdFile(stdFile = NULL)
}
\arguments{
\item{stdfile}{- std file from a TCSAM2015 model run. can be NULL.}
}
\value{
a dataframe (or NULL).
}
\description{
Function to read a TCSAM2015 std file and return a dataframe.
}
\details{
If stdFile is NULL, the user will be prompted to identify a 
TCSAM2015 model std file to read.
}

