% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeModelResultsToCSV.ObjFunValues.R
\name{writeModelResultsToCSV.ObjFunValues}
\alias{writeModelResultsToCSV.ObjFunValues}
\title{Plot all objective function values}
\usage{
writeModelResultsToCSV.ObjFunValues(repObjs, variable = "objfun",
  csv = "ModelComparisons.ObjFunValues.csv", verbose = FALSE)
}
\arguments{
\item{repObjs}{- model report object or list of report objects}

\item{variable}{- name of variable to plot}

\item{csv}{- csf file name to export dataframe to}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
dataframe with objective function components for penalties, priors, and data.
}
\description{
Function to plot all objective function values.
}

