% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeModelResultsToCSV.PopProcesses.R, R/writeModelResultsToCSV.PopQuants.R
\name{writeModelResultsToCSV.PopQuants}
\alias{writeModelResultsToCSV.PopQuants}
\title{Write population processes from model results from TCSAM2015 and rsimTCSAM model runs to a csv file.}
\usage{
writeModelResultsToCSV.PopQuants(tcsams = NULL, rsims = NULL,
  csv = "ModelResults.PopQuants", verbose = FALSE)

writeModelResultsToCSV.PopQuants(tcsams = NULL, rsims = NULL,
  csv = "ModelResults.PopQuants", verbose = FALSE)
}
\arguments{
\item{tcsams}{- single TCSAM2015 model report object, or named list of such}

\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{csv}{- base name for csv files to write to}

\item{verbose}{- flag (T/F) to print debug info}

\item{tcsams}{- single TCSAM2015 model report object, or named list of such}

\item{rsims}{- single rsimTCSAM results object, or named list of such}

\item{csv}{- base name for csv files to write to}

\item{verbose}{- flag (T/F) to print debug info}
}
\value{
nothing

nothing
}
\description{
Function to write population processes from model results from TCSAM2015 and rsimTCSAM model runs to a csv file.

Function to write population quantities from model results from TCSAM2015 and rsimTCSAM model runs to a csv file.
}
\details{
none.

none.
}

